% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.testthat.R
\name{as.expectation.logical}
\alias{as.expectation.logical}
\title{testthat function: Function to return expectation after being executed}
\usage{
as.expectation.logical(x, message, ..., srcref = NULL, info = NULL)
}
\arguments{
\item{x}{object to test for class membership}

\item{message}{\code{character} character string to be the execution message}

\item{...}{Unused, just defined inside testthat}

\item{srcref}{Only needed in very rare circumstances where you need to
forward a srcref captured elsewhere.}

\item{info}{Additional information. Included for backward compatibility
only and new expectations should not use it.}
}
\description{
In comparison to \code{testthat} this function always exports the message, even
in case of success
}
\author{
Sebastian Wolf
}
