% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_regionRef_array_master}
\alias{assign_regionRef_array_master}
\title{assign_regionRef_array_master}
\usage{
assign_regionRef_array_master(num_funcs)
}
\arguments{
\item{num_funcs}{Required length of array to store regionRef for each func}
}
\value{
R_NilValue
}
\description{
Array is not assigned on master,
rather is signalled to assign on server
}
