% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_regionRef_array_slave}
\alias{assign_regionRef_array_slave}
\title{assign_regionRef_array_slave}
\usage{
assign_regionRef_array_slave(num_funcs)
}
\arguments{
\item{num_funcs}{Required length of array to store regionRef for each func}
}
\value{
R_NilValue
}
\description{
assign_regionRef_array_slave
}
