% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{body_default_function}
\alias{body_default_function}
\title{body_default_function}
\usage{
body_default_function(func_ptr, evtID)
}
\arguments{
\item{func_ptr}{function - Pointer to function}

\item{evtID}{int - ID number for function (func_index)}
}
\value{
type(as.call) Updated function body
}
\description{
Returns instrumented body for default functions
}
