% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{body_end_cluster_function}
\alias{body_end_cluster_function}
\title{body_end_cluster_function}
\usage{
body_end_cluster_function(func_ptr, regionRef)
}
\arguments{
\item{func_ptr}{Pointer to function}

\item{regionRef}{OTF2 regionRef for function}
}
\value{
type(as.call) Updated function body
}
\description{
Returns instrumented body for functions of type end fork
}
