% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{callR}
\alias{callR}
\title{test for calling an R function}
\usage{
callR(Dmat, dvec, Amat, f)
}
\arguments{
\item{Dmat, dvec, Amat}{See \code{\link[quadprog]{solve.QP}} for explanation.}
}
\value{
a list containing the solution
}
\description{
Call the R function quadprog::solve.QP to solve a quadratic programming problem
}

