% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{clearTest}
\alias{clearTest}
\alias{clearTest,RTestCase-method}
\title{Clear Test Reporter and Test Cache of Last Test Case Execution}
\usage{
\S4method{clearTest}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\value{
(\code{\link{RTestCase-class}})
}
\description{
This method resets the test information of the current test case including slots
\code{'test.for'}, \code{'test.result'} and the cache in the \code{'tests'} test case
environment.
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
