% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{close_EvtWriterSocket_client}
\alias{close_EvtWriterSocket_client}
\title{close_EvtWriterSocket_client}
\usage{
close_EvtWriterSocket_client()
}
\value{
R_NilValue
}
\description{
close_EvtWriterSocket_client
}
