% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolveCpp}
\alias{convolveCpp}
\title{function convolve in C++ version}
\usage{
convolveCpp(a, b)
}
\arguments{
\item{a, b}{two vectors to be convolved}
}
\value{
a vector as the convolution of vectors a and b
}
\description{
The function to calculate the convolution of two vectors implemented by C++. It
plays as a performance comparison with \code{\link{convolveR}}.
}

