% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve.R
\name{convolveR}
\alias{convolveR}
\title{Function convolve in R version}
\usage{
convolveR(a, b)
}
\arguments{
\item{a, b}{two vectors to be convolved}
}
\value{
a vector as the convolution of a and b
}
\description{
The function to calculate convolution implemented by R. It comes from the famous
 example of \href{https://cran.r-project.org/doc/manuals/r-release/R-exts.html}{
 Writing R extensions}.
}
\examples{
convolveR(1:3, 2:5)
}

