% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{create_dataframe}
\alias{create_dataframe}
\title{create_dataframe}
\usage{
create_dataframe(flag_user_functions = FALSE, flag_debug = FALSE)
}
\arguments{
\item{flag_user_functions}{Boolean - TRUE to include user functions in dataframe}

\item{flag_debug}{Boolean - Enabled debug state,emts}
}
\value{
Dataframe containing each function with information on function name,
package, count of function calls, total time spent in function
}
\description{
See return
}
