% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{create_eventlog}
\alias{create_eventlog}
\title{create_eventlog
See return}
\usage{
create_eventlog()
}
\value{
Dataframe for event logging
}
\description{
create_eventlog
See return
}
