% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{define_otf2_event_client}
\alias{define_otf2_event_client}
\title{define_otf2_event_client}
\usage{
define_otf2_event_client(func_name)
}
\arguments{
\item{func_name}{Name of function to create event for}
}
\value{
regionRef regionRef for use when logging events
}
\description{
define_otf2_event_client
}
