% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{define_otf2_regionRef_client}
\alias{define_otf2_regionRef_client}
\title{define_otf2_regionRef_client}
\usage{
define_otf2_regionRef_client(func_name, func_index)
}
\arguments{
\item{func_name}{Name of function to create event for}

\item{func_index}{Global index of function in R namespace}
}
\value{
regionRef regionRef for use when logging events
}
\description{
define_otf2_regionRef_client
}
