% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Title}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a data frame}

\item{varx}{a string name for variabel x dataframe called df}

\item{vary}{a string name for variabel y dataframe called df}
}
\value{
A plot with diamond markers of \code{varx} versus \code{vary}
}
\description{
Title
}
\examples{
df.test <- data.frame(x = rnorm(10),
                      y = rnorm(10))

diamond_plot(df.test, 'x', 'y')
}
