% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{disable_instrumentation}
\alias{disable_instrumentation}
\title{disable_instrumentation
Disable instrumentation}
\usage{
disable_instrumentation()
}
\description{
disable_instrumentation
Disable instrumentation
}
