% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{enable_instrumentation}
\alias{enable_instrumentation}
\title{enable_instrumentation
Enable instrumentation and reset function depth}
\usage{
enable_instrumentation()
}
\description{
enable_instrumentation
Enable instrumentation and reset function depth
}
