% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epoch_time_client}
\alias{epoch_time_client}
\title{Send current timestamp to server. Only used for epoch_start and epoch_end}
\usage{
epoch_time_client()
}
\value{
R_NilValue
}
\description{
Send current timestamp to server. Only used for epoch_start and epoch_end
}
