% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evtWriter_MeasurementOnOff}
\alias{evtWriter_MeasurementOnOff}
\title{Enable or disable event measurement}
\usage{
evtWriter_MeasurementOnOff(measurementMode)
}
\arguments{
\item{measurementMode}{True to enable, else disable}
}
\value{
R_NilValue
}
\description{
Enable or disable event measurement
}
