% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evtWriter_proc_client}
\alias{evtWriter_proc_client}
\title{evtWriter_proc_client}
\usage{
evtWriter_proc_client(proc_init)
}
\arguments{
\item{proc_init}{True if initializing new proc, else false if finalizing proc}
}
\value{
R_NilValue
}
\description{
Signal to server to send regionRef array to new procs
}
