% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evtWriter_Write_client}
\alias{evtWriter_Write_client}
\title{Write event to evt_writer}
\usage{
evtWriter_Write_client(func_index, event_type)
}
\arguments{
\item{func_index}{Function index in global list}

\item{event_type}{True for enter, False for leave region}
}
\value{
R_NilValue
}
\description{
Write event to evt_writer
}
