% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.examples.R
\name{example_data_frame}
\alias{example_data_frame}
\title{Function changing a data frame by adding a column}
\usage{
example_data_frame(data = data.frame(x = c(1, 2), y = c(1, 2)),
  mult = 1)
}
\arguments{
\item{data}{(\code{data.frame}) Any data frame with numeric values}

\item{mult}{(\code{numeric}) Any numeric value (length == 1)}
}
\value{
A data.frame with an additional column sum that
  is the rowwise sum multiplied by \code{mult}
}
\description{
Function changing a data frame by adding a column
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
