% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.examples.R
\name{example_image}
\alias{example_image}
\title{Function returning the Roche logo as an image at tempdir}
\usage{
example_image(name = "Roche_logo.png")
}
\arguments{
\item{name}{\code{character} The name of the output image}
}
\value{
The file path to a temporar file with the given
	name that will contain the Roche_Logo.png that
 comes with RTest/images
}
\description{
Function returning the Roche logo as an image at tempdir
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
