% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.examples.R
\name{example_list}
\alias{example_list}
\title{Function returning a list with three values}
\usage{
example_list(name_1 = "NAME1", value_2 = 1)
}
\arguments{
\item{name_1}{(\code{character}) Name of the first list element}

\item{value_2}{(\code{numeric}) Value of the second list element}
}
\value{
A list with three elements, a generic data frame
 inside the element \code{data.frame} a list element with
 the value "VALUE1" inside the element with name of
 parameter \code{name_1} and an item with the name "NAME2"
 and the value of \code{value_2} inside.
}
\description{
Function returning a list with three values
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
