% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.examples.R
\name{example_variable}
\alias{example_variable}
\title{Function returning relative difference of X and Y}
\usage{
example_variable(x = 1.2, y = 1)
}
\arguments{
\item{x}{(\code{numeric}) X-value}

\item{y}{(\code{numeric}) Y-value}
}
\value{
\code{(X-Y)/(X)}
}
\description{
Function returning relative difference of X and Y
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
