% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.examples.R
\name{example_vector}
\alias{example_vector}
\title{Function returning a character vector of length "rep"}
\usage{
example_vector(rep = 5)
}
\arguments{
\item{rep}{(\code{numeric}) Number of repetitions}
}
\value{
character vector containing \code{rep} times
  the word "RTest"
}
\description{
Function returning a character vector of length "rep"
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
