% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.testthat.R
\name{expect_less_than}
\alias{expect_less_than}
\alias{expect_more_than}
\alias{expect_lt}
\alias{expect_less_more}
\alias{expect_gt}
\title{Expect less or more}
\usage{
expect_less_than(...)

expect_more_than(...)

expect_lt(...)

expect_gt(...)
}
\arguments{
\item{...}{Any parameter sent to expect_gt or expect_lt of testthat}
}
\description{
Expect less or more
}
