% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.testthat.R
\name{expect_silent_RTest}
\alias{expect_silent_RTest}
\title{Expect a function call to run silent}
\usage{
expect_silent_RTest(object)
}
\arguments{
\item{object}{executable function call}
}
\description{
In case the function call is not silent, a message including all outputs, messages, warnings
is given.
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
