% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.testthat.R
\name{expect_testthat}
\alias{expect_testthat}
\title{testthat expect fun}
\usage{
expect_testthat(ok, failure_message, info = NULL, srcref = NULL)
}
\arguments{
\item{ok}{(\code{logcal}) Whether code was ok}

\item{failure_message}{\code{character} What error cane out}

\item{info}{(\code{list}) List of infos given for execution}

\item{srcref}{(\code{call}) Call that was executed}
}
\description{
testthat expect fun
}
