% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{finalize_Archive}
\alias{finalize_Archive}
\title{Close static otf2 {archive, evt_writer} objs}
\usage{
finalize_Archive()
}
\value{
R_NilValue
}
\description{
Close static otf2 {archive, evt_writer} objs
}
