% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{finalize_GlobalDefWriter}
\alias{finalize_GlobalDefWriter}
\title{Finalize static otf2 {globaldefwriter} obj
Write clock information before ending tracing}
\usage{
finalize_GlobalDefWriter()
}
\value{
R_NilValue
}
\description{
Finalize static otf2 {globaldefwriter} obj
Write clock information before ending tracing
}
