% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{finalize_GlobalDefWriter_client}
\alias{finalize_GlobalDefWriter_client}
\title{finalize_GlobalDefWriter_client}
\usage{
finalize_GlobalDefWriter_client()
}
\value{
RNilValue
}
\description{
finalize_GlobalDefWriter_client
}
