% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{finalize_sync_client}
\alias{finalize_sync_client}
\title{finalize_sync_client}
\usage{
finalize_sync_client()
}
\value{
R_NilValue
}
\description{
Send signal to server to stop collecting event information
}
