% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.test-adapter.R
\name{get_existence_of_fun}
\alias{get_existence_of_fun}
\title{Find out if a function is available}
\usage{
get_existence_of_fun(function_name, package)
}
\arguments{
\item{function_name}{(\code{character}) The name of the function to look up}

\item{package}{(\code{character}) The name of the package where this function
might be hidden (not exported)}
}
\value{
"global" if it is available, "package" if it's available within the package
 or an Error if it is not available at all.
}
\description{
This function checks if a function name or method name is available in the global
namespace or the desired package namespace
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
