% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{get_fork_wrapper_expression}
\alias{get_fork_wrapper_expression}
\title{get_fork_wrapper_expression}
\usage{
get_fork_wrapper_expression()
}
\description{
Returns wrapper expression
Split between start and end because makeForkCluster contains line \verb{on.exit(<cmd>)}, without
arg \code{add=TRUE}. Overwrites my commands as a result. Also have to manually add final return line
but this could be generalized better by splitting original function body
}
