% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_exception_list.R
\name{get_function_exception_list}
\alias{get_function_exception_list}
\title{get_function_exception_list()}
\usage{
get_function_exception_list()
}
\value{
function_exception_list Object[] List of all functions to not instrument
}
\description{
Get function list to not insturment (mostly due to recursion problems in wrapper)
Note: This is a hacky fix for many smaller problems
}
