% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{get_function_index}
\alias{get_function_index}
\title{get_function_index}
\usage{
get_function_index(func_list, func_ptr, func_name)
}
\arguments{
\item{func_list}{Object[] - Array of functions}

\item{func_ptr}{Object - Function}

\item{func_name}{String - Name of function}
}
\value{
index Integer - for func_name in func_list, else NULL if not found
}
\description{
Get function index from func_list
}
