% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_exception_list.R
\name{get_function_methods}
\alias{get_function_methods}
\title{get_function_methods}
\usage{
get_function_methods(function_list)
}
\arguments{
\item{function_list}{Object[] List of function pointers}
}
\value{
function_method_list Object[] List of all method functions for function_list
}
\description{
Get all methods of given function list
}
