% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{get_new_function_body}
\alias{get_new_function_body}
\title{get_new_function_body}
\usage{
get_new_function_body(func_ptr, func_name, evtID)
}
\arguments{
\item{func_ptr}{function - Pointer to function}

\item{func_name}{Char[] - Name of function, only used for debugging}

\item{evtID}{int - ID number for function (func_index)}
}
\value{
type(as.call) Updated function body
}
\description{
Returns instrumented body for function type {fork, end fork, default}
}
