% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{get_psock_function_list}
\alias{get_psock_function_list}
\title{get_psock_function_list}
\usage{
get_psock_function_list()
}
\description{
Returns list of known functions which R uses to fork procs
}
