% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{get_psock_wrapper_expression}
\alias{get_psock_wrapper_expression}
\title{get_psock_wrapper_expression}
\usage{
get_psock_wrapper_expression()
}
\value{
Returns wrapper expression
}
\description{
List containing wrapper expression to insert at start and end of makePSOCKcluster and similar functions
}
