% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_regionRef_array_master}
\alias{get_regionRef_array_master}
\title{get_regionRef_array_master}
\usage{
get_regionRef_array_master(nprocs)
}
\arguments{
\item{nprocs}{Number of new procs to update}
}
\value{
R_NilValue
}
\description{
Signal to server to send regionRef array to new procs
}
