% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_regionRef_array_slave}
\alias{get_regionRef_array_slave}
\title{get_regionRef_array_slave}
\usage{
get_regionRef_array_slave(num_funcs)
}
\arguments{
\item{num_funcs}{Total number of functions in R namespace}
}
\value{
R_NilValue
}
\description{
Requests regionRef array from logger proc
}
