% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_regionRef_from_array_slave}
\alias{get_regionRef_from_array_slave}
\title{get_regionRef_from_array_slave}
\usage{
get_regionRef_from_array_slave(func_index)
}
\arguments{
\item{func_index}{Index of function to fet regionRef for}
}
\value{
regionRef
}
\description{
get_regionRef_from_array_slave
}
