% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{get_user_function_list}
\alias{get_user_function_list}
\title{get_user_function_list}
\usage{
get_user_function_list(flag_debug = FALSE)
}
\arguments{
\item{flag_debug}{Boolean - Enable debug output}
}
\value{
user_func_list Function[] - List of user-defined functions
}
\description{
Get all user-defined functions in global_env
}
