% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{getExecDetails.html}
\alias{getExecDetails.html}
\alias{getExecDetails.html,RTestCase-method}
\title{Create Detailed HTML Summary of the Last Execution of the Test Case}
\usage{
\S4method{getExecDetails.html}{RTestCase}(object,
  report.onlyFailed = FALSE)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}

\item{report.onlyFailed}{(\code{logical}) Report only failed exceptions (TRUE) or
all exceptions (FALSE, default).}
}
\value{
(\code{character})
}
\description{
Generates a detailed summary of the last test case execution for the overall test report
including all tested expectations.
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
