% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{getExecStates}
\alias{getExecStates}
\alias{getExecStates,RTestCollection-method}
\title{Execution Statues of TCs}
\usage{
\S4method{getExecStates}{RTestCollection}(object, test.TCs = NULL)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object.}

\item{test.TCs}{(\code{character}) Vector with the TCs to be executed or NULL if all
all TCs of the collection should be tested.}
}
\value{
\code{\link{character}} Vector with test result ('failed' or 'passed') for all TCs.
}
\description{
Execution Statues of TCs
}
\seealso{
\code{\link{RTestCollection-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
