% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{getExecSummary.html}
\alias{getExecSummary.html}
\alias{getExecSummary.html,RTestCase-method}
\title{Create HTML Summary of the Last Execution of the Test Case}
\usage{
\S4method{getExecSummary.html}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\value{
(\code{character})
}
\description{
Generates a summary of the last test case execution for the overall test report.
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
