% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{getExecSummary}
\alias{getExecSummary}
\alias{getExecSummary,RTestCase-method}
\title{Summary of the Last Execution of the Test Case}
\usage{
\S4method{getExecSummary}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\value{
(\code{list})
}
\description{
Summary of the Last Execution of the Test Case
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
