% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getfun}
\alias{getfun}
\title{function to derive external package functionalities}
\usage{
getfun(x)
}
\arguments{
\item{x}{\code{character} package :: function string}
}
\value{
functionality of the wanted function

found at 
https://stackoverflow.com/questions/38983179/do-call-a-function-in-r-without-loading-the-package
}
\description{
function to derive external package functionalities
}
