% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{getRTM}
\alias{getRTM}
\alias{getRTM,RTestCollection-method}
\title{Get RTM for all executed test cases}
\usage{
\S4method{getRTM}{RTestCollection}(object, test.TCs = NULL, ...)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object.}

\item{test.TCs}{(\code{character}) Vector with the TCs to be executed or NULL if all
all TCs of the collection should be tested.}

\item{...}{Additional arguments passed to \code{\link{getRTMInfos}}.}
}
\value{
(\code{data.frame}) The RTM as \code{data.frame} table object.
}
\description{
This method returns the requirement traceability matrix (RTM) for all imported test cases.
}
\seealso{
\code{\link{RTestCollection-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
