% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{getSynopsis}
\alias{getSynopsis}
\alias{getSynopsis,RTestCase-method}
\title{Get Synopsis of the Test Case}
\usage{
\S4method{getSynopsis}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\value{
(\code{character})
}
\description{
Get Synopsis of the Test Case
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
