% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{getTC}
\alias{getTC}
\alias{getTC,RTestCollection-method}
\title{Return Imported Test Case}
\usage{
\S4method{getTC}{RTestCollection}(object, tc.id)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object}

\item{tc.id}{(\code{character}) TestCase ID}
}
\description{
Return Imported Test Case
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
