% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{getTestFor}
\alias{getTestFor}
\alias{getTestFor,RTestCase-method}
\title{Get For of Last Execution of the Test Case}
\usage{
\S4method{getTestFor}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\value{
(\code{ANY})
}
\description{
Get For of Last Execution of the Test Case
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
