% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{getValidTCs}
\alias{getValidTCs}
\alias{getValidTCs,RTestCollection-method}
\title{Get a List of TCs}
\usage{
\S4method{getValidTCs}{RTestCollection}(object, test.TCs = NULL)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object.}

\item{test.TCs}{(\code{character}) Vector with the TCs to be executed or NULL if all
all TCs of the collection should be tested.}
}
\value{
(\code{character}) Only the valid TCs
}
\description{
Get a List of TCs
}
\seealso{
\code{\link{RTestCollection-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
