% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{getXMLRoot}
\alias{getXMLRoot}
\alias{getXMLRoot,RTestCase-method}
\title{Get the XML Root of the Test Case}
\usage{
\S4method{getXMLRoot}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\value{
(\code{XMLNode})
}
\description{
Get the XML Root of the Test Case
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
