% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{getXMLSourcePath}
\alias{getXMLSourcePath}
\alias{getXMLSourcePath,RTestCase-method}
\title{Get Path to the XML Input File of the Test Case}
\usage{
\S4method{getXMLSourcePath}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\value{
(\code{character})
}
\description{
Get Path to the XML Input File of the Test Case
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
