% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{globalDefWriter_WriteLocation}
\alias{globalDefWriter_WriteLocation}
\title{Write a definition for the location to the global definition writer.}
\usage{
globalDefWriter_WriteLocation(stringRef_name)
}
\arguments{
\item{stringRef_name}{Name to be associated with SystemTreeNode (eg MyHost)}
}
\value{
R_NilValue
}
\description{
Write a definition for the location to the global definition writer.
}
