% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{globalDefWriter_WriteRegion}
\alias{globalDefWriter_WriteRegion}
\title{globalDefWriter_WriteRegion
Define new region description in global writer}
\usage{
globalDefWriter_WriteRegion(stringRef_RegionName)
}
\arguments{
\item{stringRef_RegionName}{Name to be associated with region}
}
\value{
regionRef id/index for string
}
\description{
globalDefWriter_WriteRegion
Define new region description in global writer
}
