% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{globalDefWriter_WriteString}
\alias{globalDefWriter_WriteString}
\title{Define new id-value pair in globaldefwriter}
\usage{
globalDefWriter_WriteString(stringRef, stringRefValue)
}
\arguments{
\item{stringRef}{id/index for stringRef}

\item{stringRefValue}{String assigned to given id}
}
\value{
NUM_STRINGREF
}
\description{
Define new id-value pair in globaldefwriter
}
