% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{globalDefWriter_WriteSystemTreeNode}
\alias{globalDefWriter_WriteSystemTreeNode}
\title{Write the system tree including a definition for the location group to the global definition writer.}
\usage{
globalDefWriter_WriteSystemTreeNode(stringRef_name, stringRef_class)
}
\arguments{
\item{stringRef_name}{Name to be associated with SystemTreeNode (eg MyHost)}

\item{stringRef_class}{Class to be associated with SystemTreeNode (eg node)}
}
\value{
R_NilValue
}
\description{
Write the system tree including a definition for the location group to the global definition writer.
}
