% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{htmlify_string}
\alias{htmlify_string}
\title{function to make strings xml and html compatible}
\usage{
htmlify_string(input_string)
}
\arguments{
\item{input_string}{(\code{character}) A simple character string}
}
\value{
(\code{character}) The same string but incompatible characters
are exchanged by HTML Name characters such as &amp; for &
}
\description{
function to make strings xml and html compatible
}
\examples{

input_string <- "<5"

stopifnot(htmlify_string(input_string)=="&lt;5")

}
\author{
Sebastian Wolf <sebastian.wolf.sw1@roche.com>
}
