% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{importTC}
\alias{importTC}
\alias{importTC,RTestCollection-method}
\title{Import Test Case from XML File}
\usage{
\S4method{importTC}{RTestCollection}(object, xml.fPath)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object}

\item{xml.fPath}{(\code{character}) Path to XML definition file}
}
\description{
This function imports a test case XML definition and adds it to the test collection.
}
\seealso{
\code{\link{RTestCollection-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
