% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{importTCsFromDir}
\alias{importTCsFromDir}
\alias{importTCsFromDir,RTestCollection-method}
\title{Import all Test Cases from XML Files of a Directory}
\usage{
\S4method{importTCsFromDir}{RTestCollection}(object, xml.dPath,
  f.pattern = "\\\\.xml$", f.igncase = FALSE)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object}

\item{xml.dPath}{(\code{character}) Path to directory containing the XML files}

\item{f.pattern}{(\code{character}) An optional regular expression to search the input
directory. Only file names which match the regular expression will be
returned (passed as argument to \code{\link[base]{list.files}}).}

\item{f.igncase}{(\code{logical}) Should pattern-matching be case-insensitive or nt
(passed as argument to \code{\link[base]{list.files}}).}
}
\description{
This function imports a test case XML definition and adds it to the test collection.
}
\seealso{
\code{\link{RTestCollection-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
