% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_Archive}
\alias{init_Archive}
\title{Initialize static otf2 {archive} objs}
\usage{
init_Archive(archivePath = "./rTrace", archiveName = "rTrace")
}
\arguments{
\item{archivePath}{Path to the archive i.e. the directory where the anchor file is located.}

\item{archiveName}{Name of the archive. It is used to generate sub paths e.g. "archiveName.otf2"}
}
\value{
R_NilValue
}
\description{
Initialize static otf2 {archive} objs
}
