% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_GlobalDefWriter}
\alias{init_GlobalDefWriter}
\title{Init static otf2 {globaldefwriter} obj}
\usage{
init_GlobalDefWriter()
}
\value{
R_NilValue
}
\description{
Init static otf2 {globaldefwriter} obj
}
