% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_otf2_logger}
\alias{init_otf2_logger}
\title{Fork and initialize zeromq sockets for writing globalDef definitions}
\usage{
init_otf2_logger(
  max_nprocs,
  archivePath = "./rTrace",
  archiveName = "rTrace",
  ports = as.numeric(c(5556, 5557)),
  flag_print_pids = FALSE
)
}
\arguments{
\item{max_nprocs}{Maximum number of R processes (ie evtWriters required)}

\item{archivePath}{Path to otf2 archive}

\item{archiveName}{Name of otf2 archive}

\item{ports}{Port numbers for zmq sockets (must be length <NUM_PORTS)}

\item{flag_print_pids}{True to print pids of parent and child procs}
}
\value{
<0 if error, 0 if R master, else >0 if child
}
\description{
Fork and initialize zeromq sockets for writing globalDef definitions
}
