% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{initializeTests}
\alias{initializeTests}
\alias{initializeTests,RTestCase-method}
\title{Initialize the Test Slot for a Test Case.}
\usage{
\S4method{initializeTests}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\description{
This method initializes the slots '\code{tests}', '\code{test.for}' and '\code{test.result}' of
a object of class '\code{RTestCase}'. See description of \code{\link{RTestCase-class}} for
further information.
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
