% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_ll.R
\name{instrument_all_functions}
\alias{instrument_all_functions}
\title{instrument_all_functions}
\usage{
instrument_all_functions(
  package_list = NULL,
  flag_user_functions = TRUE,
  flag_print_progress = TRUE,
  flag_debug = FALSE,
  flag_slave_proc = FALSE
)
}
\arguments{
\item{package_list}{String[] - Array of package names to instrument, if none instrument all packages}

\item{flag_user_functions}{Boolean - TRUE if also flagging user functions}

\item{flag_print_progress}{Boolean - Enable for package by package progress statements}

\item{flag_debug}{Boolean - Enable debug statements}

\item{flag_slave_proc}{Boolean - Enable if running on slave proc}
}
\description{
Instrument all functions
}
