% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_hl.R
\name{instrumentation_debug}
\alias{instrumentation_debug}
\title{instrumentation_debug}
\usage{
instrumentation_debug(
  print_func_indexes = pkg.env$PRINT_FUNC_INDEXES,
  max_function_depth = pkg.env$MAX_FUNCTION_DEPTH,
  unlock_env = pkg.env$UNLOCK_ENVS
)
}
\arguments{
\item{print_func_indexes}{info}

\item{max_function_depth}{info}

\item{unlock_env}{info}
}
\description{
Enable certain debug features
}
