% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_hl.R
\name{instrumentation_disable}
\alias{instrumentation_disable}
\title{instrumentation_disable}
\usage{
instrumentation_disable(flag_check_depth = T, flag_update_measurement = TRUE)
}
\arguments{
\item{flag_check_depth}{Boolean - Intended for developers, suppress depth warning if false}

\item{flag_update_measurement}{Boolean - Intended for developers, update measurement mode}
}
\description{
Disable instrumentation
}
