% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_hl.R
\name{instrumentation_enable}
\alias{instrumentation_enable}
\title{instrumentation_enable}
\usage{
instrumentation_enable(flag_ignore_depth = FALSE)
}
\arguments{
\item{flag_ignore_depth}{Boolean - Intended for developers, suppress depth warning}
}
\description{
Enable instrumentation and reset function depth
}
