% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{instrumentation_finalize}
\alias{instrumentation_finalize}
\title{instrumentation_finalize
Close otf2 objs for instrumentation}
\usage{
instrumentation_finalize()
}
\description{
instrumentation_finalize
Close otf2 objs for instrumentation
}
