% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_hl.R
\name{instrumentation_wrapper}
\alias{instrumentation_wrapper}
\title{instrumentation_wrapper}
\usage{
instrumentation_wrapper(func, ...)
}
\arguments{
\item{func}{Object - Function to call}

\item{...}{Args - Function args}
}
\description{
Simple function to provider wrapper for instrumenting a single function call.
Ideal use if program contained in main(). Not intended to be used multiple times in one script
}
