% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{is_cluster_function}
\alias{is_cluster_function}
\title{is_cluster_function}
\usage{
is_cluster_function(func_ptr, fork_cluster = F)
}
\arguments{
\item{func_ptr}{Function to check}

\item{fork_cluster}{True if checking for fork function, else check for psock}
}
\value{
Boolean - true if fork/psock function, else false
}
\description{
Returns bool for if function listed as for type of {fork, psock} function
}
