% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{is_end_fork_function}
\alias{is_end_fork_function}
\title{is_end_fork_function}
\usage{
is_end_fork_function(func_ptr)
}
\arguments{
\item{func_ptr}{Function to check}
}
\value{
Boolean - true if end fork function, else false
}
\description{
Returns bool for if function listed as for type of end fork function
}
