% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{is_instrumentation_enabled}
\alias{is_instrumentation_enabled}
\title{is_instrumentation_enabled
Return current instrumentation status}
\usage{
is_instrumentation_enabled()
}
\value{
BOOLEAN - Instrumentation status
}
\description{
is_instrumentation_enabled
Return current instrumentation status
}
