% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_hl.R
\name{is_instrumentation_init}
\alias{is_instrumentation_init}
\title{is_instrumentation_init}
\usage{
is_instrumentation_init()
}
\value{
TRUE if init, else FALSE
}
\description{
Error catching function to ensure instrumentation_init() has been called
}
