% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{isChrConsistent}
\alias{isChrConsistent}
\title{Verify chromosome names}
\usage{
isChrConsistent(pacds, obj, col=NULL, allin=FALSE)
}
\arguments{
\item{pacds}{a PACdataset with @anno$chr, or df/matrix with chr column, or vector}

\item{obj}{BSgenome, or FaFile, or data.frame/matrix or vector}

\item{col}{the colname of chr in obj[if obj is df or matrix]}

\item{allin}{pacds$chr should all in obj's chr, otherwise intersect is allowed}
}
\value{
TRUE (consistent or intersect) or FALSE
}
\description{
Verify chromosome names
}
\examples{
isChrConsistent(PACds, obj=bsgenome, allin=TRUE)
isChrConsistent(PACds, obj=FaFile(aFastaFile), allin=TRUE)
isChrConsistent(PACds, obj=unique(PACds@anno$chr), allin=TRUE)
isChrConsistent(PACds, obj=c('x','a'), allin=TRUE)
isChrConsistent(PACds, obj=d, allin=TRUE)
isChrConsistent(PACds, obj=d, col='strand',allin=TRUE)
isChrConsistent(c('1','2'), obj=c('chr1','chr2'),allin=TRUE)
}
