% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utils.R
\name{lock_envs}
\alias{lock_envs}
\title{lock_envs}
\usage{
lock_envs(package_name)
}
\arguments{
\item{package_name}{String - name of package}
}
\description{
Lock package- and name-spaces to change function definitions in packages
Counterfunction for \code{unlock_envs}
}
