% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_add_abcn}
\alias{mult_add_abcn}
\title{S_abcn}
\usage{
mult_add_abcn(a, b, c, n)
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{c}{c}

\item{n}{Number of loops}
}
\value{
R_NilValue
}
\description{
S_abcn
}
