% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_add_n}
\alias{mult_add_n}
\title{Simple multiply+add operation for timing}
\usage{
mult_add_n(n)
}
\arguments{
\item{n}{Number of loops}
}
\value{
R_NilValue
}
\description{
Simple multiply+add operation for timing
}
