% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{open_EvtWriterSocket_client}
\alias{open_EvtWriterSocket_client}
\title{open_EvtWriterSocket_client}
\usage{
open_EvtWriterSocket_client()
}
\value{
R_NilValue
}
\description{
open_EvtWriterSocket_client
}
