% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{orderBeta}
\alias{orderBeta}
\title{Generating a partial ordered matrix of beta}
\usage{
orderBeta(p, q, theta, ref = 1L)
}
\arguments{
\item{p, q}{Dimensions of the beta matrix}

\item{theta}{Probability that adjacent beta's have the same value}

\item{ref}{Whether set the first beta to 0}
}
\value{
A matrix of beta's
}
\description{
This function generates a matrix of beta with size p*q, with a partial order in both direction. A parameter
theta control the possibility of fusion of adjecent beta.
}

