% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{otf2_handle_proc}
\alias{otf2_handle_proc}
\title{otf2_handle_proc}
\usage{
otf2_handle_proc(is_init)
}
\arguments{
\item{is_init}{True if init proc, else false if finalizing proc}
}
\value{
R_NilValue
}
\description{
Signal to server to send regionRef array to new procs
}
