% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\docType{class}
\name{PACdataset-class}
\alias{PACdataset-class}
\alias{PACdataset}
\title{PACdataset object}
\description{
PACdataset is an S4 class to represent a poly(A) site list.
PACs are stored as PACdataset for filtering, statistics, and other operations.
}
\section{Slots}{

\describe{
\item{\code{counts}}{a data frame denoting the counts of PAs. Each row is one PA, each column is one experiment.}

\item{\code{colData}}{sample defination. Each row is one experiment, and the column is the sample group.
The number of rows is the same as the column number of the counts slot.}

\item{\code{anno}}{detailed annotation information for PAs, with the same row number as the counts slot.}

\item{\code{supp}}{a list storing additional data, like stopCodon.}
}}

\seealso{
Other PACdataset functions: 
\code{\link{rbind}()},
\code{\link{subsetPACds}()}
}
\concept{PACdataset functions}
