% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_md5}
\alias{package_md5}
\title{Function to derive an md5 hash of a package in a current session}
\usage{
package_md5(package)
}
\arguments{
\item{package}{(\code{character}) Name of the package to be scanned}
}
\value{
\code{Namespace: } md5 hash or \code{tar: } md5 hash. Dependent
on whether the package can be loaded out of its current Namespace (Namespace) or
whether the installed binaries have to be scanned (tar) a different
md5 hash is given. Namespace packages where normally loaded using \code{library}
calls
}
\description{
Function to derive an md5 hash of a package in a current session
}
\examples{

package_md5("testthat")

}
\author{
Sebastian Wolf \email{sebastian.wolf.sw1@roche.com}
}
