% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAtest.R
\name{PCA.analysis}
\alias{PCA.analysis}
\title{PCA.analysis}
\usage{
PCA.analysis(data.omics, lambda = 1)
}
\arguments{
\item{data.omics}{data.omics is input data frame or matrix}
}
\value{
a dataframe
}
\description{
This function is for PCA analysis to transform the methylation data.
}
\details{
LEMC: A likelihood-based method for estimating methylation-induced correlation among human complex traits

you can use this function to calculate the eigen value and eigen matrix of data.
}
\examples{
\dontrun{
## a simulated DNA methylation data
load('data/data.example.Rdata')
library(Morpho)
PCA.data <- PCA.analysis(data.omics)
PCA.data
}
}
