% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTest.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Package level environment
Used for storing package level variables (flags eg INSTRUMENTATION_ENABLED, and counters eg PROFILE_EVENTLOG)
\itemize{
\item INSTRUMENTATION_ENABLED
\item FLAG_INSTRUMENT_ALL
\item FLAG_INSTRUMENT_USER_FUNCTIONS
\item MAX_FUNCTION_DEPTH
\item FUNCTION_DEPTH
\item UNLOCK_ENVS
\item PROFILE_EVENTLOG
\item PROFILE_EVENTLOG_NROWS
\item PROFILE_INSTRUMENTATION_DF
\item PRINT_INSTRUMENTS
\item PRINT_SKIPS
\item PRINT_FUNC_INDEXES
}}
\format{
An object of class \code{environment} of length 0.
}
\usage{
pkg.env
}
\description{
Package level environment
Used for storing package level variables (flags eg INSTRUMENTATION_ENABLED, and counters eg PROFILE_EVENTLOG)
\itemize{
\item INSTRUMENTATION_ENABLED
\item FLAG_INSTRUMENT_ALL
\item FLAG_INSTRUMENT_USER_FUNCTIONS
\item MAX_FUNCTION_DEPTH
\item FUNCTION_DEPTH
\item UNLOCK_ENVS
\item PROFILE_EVENTLOG
\item PROFILE_EVENTLOG_NROWS
\item PROFILE_INSTRUMENTATION_DF
\item PRINT_INSTRUMENTS
\item PRINT_SKIPS
\item PRINT_FUNC_INDEXES
}
}
\keyword{datasets}
