% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{png2base64}
\alias{png2base64}
\title{Converts a PNG File Into a Base64 String for Using IT as Inline Image in HTML Files}
\usage{
png2base64(file, img.returnAsTag = FALSE, img.title = "image",
  img.width = NULL)
}
\arguments{
\item{file}{(\code{character}) Path to PNG image.}

\item{img.returnAsTag}{(\code{boolean})    TRUE, img is returned as HTML \code{img}-tag; FALSE,
raw image base64 content is returned.}

\item{img.title}{(\code{character})    Title of the HTML \code{img}-tag.}

\item{img.width}{(\code{character})    With for HTML \code{img}-tag.}
}
\value{
See paramter \code{img.returnAsTag}.
}
\description{
This function creates a base64 string of a PNG  (e.g. png) directly into HTML via
the data function.
}
\details{
The input file is read and a base64 string encoded The returned file is the value of the 
\code{img} attribute \code{src}.
}
\author{
Matthias Pfeifer <matthias.pfeifer@roche.com>
}
