% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{posdefmat}
\alias{posdefmat}
\title{function to generate random positive definite matrice}
\usage{
posdefmat(n, ev = runif(n, 0, 10))
}
\arguments{
\item{n}{size of the matrix}
}
\description{
function to generate random positive definite matrice
}

