% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{print_dataframe}
\alias{print_dataframe}
\title{print_instrumentation}
\usage{
print_dataframe(
  df = pkg.env$PROFILE_INSTRUMENTATION_DF,
  flag_reduced = TRUE,
  flag_debug = TRUE
)
}
\arguments{
\item{df}{Dataframe - Created by create_dataframe() for collecting tracing info}

\item{flag_reduced}{Boolean - True to display only non-zero count functions, else display all}

\item{flag_debug}{Boolean - Enable debug header}
}
\description{
Print table of reduce dataframe
}
