% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{print_eventlog}
\alias{print_eventlog}
\title{print_eventlog
Print global event log}
\usage{
print_eventlog(df = pkg.env$PROFILE_EVENTLOG)
}
\arguments{
\item{df}{\itemize{
\item Dataframe object for eventlog
}}
}
\description{
print_eventlog
Print global event log
}
