% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_create_dataframe}
\alias{profile_create_dataframe}
\title{profile_create_dataframe
See return}
\usage{
profile_create_dataframe(flag_debug = FALSE)
}
\arguments{
\item{flag_debug}{Boolean - Enabled debug state,emts}
}
\value{
Dataframe containing each function with information on function name,
package, count of function calls, total time spent in function
}
\description{
profile_create_dataframe
See return
}
