% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_create_eventlog}
\alias{profile_create_eventlog}
\title{profile_create_eventlog
See return}
\usage{
profile_create_eventlog()
}
\value{
Dataframe for event logging
}
\description{
profile_create_eventlog
See return
}
