% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_create_otf2_event}
\alias{profile_create_otf2_event}
\title{profile_create_otf2_event
Creates stringRef and regionRef for func_name}
\usage{
profile_create_otf2_event(func_name)
}
\arguments{
\item{func_name}{String - Name of function}
}
\value{
regionRef Int - Index of stringRef for function
}
\description{
profile_create_otf2_event
Creates stringRef and regionRef for func_name
}
