% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_enable_instrumentation}
\alias{profile_enable_instrumentation}
\title{profile_enable_instrumentation
Enable instrumentation and reset function depth}
\usage{
profile_enable_instrumentation()
}
\description{
profile_enable_instrumentation
Enable instrumentation and reset function depth
}
