% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_event_create}
\alias{profile_event_create}
\title{profile_event_create
Create new start or end event in global eventlog}
\usage{
profile_event_create(region, event_type, timestamp)
}
\arguments{
\item{region}{Integer - Index of event/function in INSTRUMENTATION_DF}

\item{event_type}{Boolean - TRUE for start, FALSE for end event}

\item{timestamp}{numeric - Time of event occurence}
}
\description{
profile_event_create
Create new start or end event in global eventlog
}
