% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_get_user_function_list}
\alias{profile_get_user_function_list}
\title{profile_get_user_function_list
Get all user-defined functions in global_env}
\usage{
profile_get_user_function_list(flag_debug = FALSE)
}
\arguments{
\item{flag_debug}{Boolean - Enable debug output}
}
\value{
user_func_list Function[] - List of user-defined functions
}
\description{
profile_get_user_function_list
Get all user-defined functions in global_env
}
