% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_instrument_all_functions}
\alias{profile_instrument_all_functions}
\title{profile_instrument_all_functions
Instrument all functions}
\usage{
profile_instrument_all_functions(
  package_list = NULL,
  flag_user_functions = TRUE,
  flag_debug = FALSE
)
}
\arguments{
\item{package_list}{String[] - Array of package names to instrument, if none instrument all packages}

\item{flag_user_functions}{Boolean - True if also flagging user functions}

\item{flag_debug}{Boolean - Enable debug statements}
}
\description{
profile_instrument_all_functions
Instrument all functions
}
