% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_instrument_user_functions}
\alias{profile_instrument_user_functions}
\title{profile_instrument_user_functions
Instrument user functions}
\usage{
profile_instrument_user_functions(flag_debug = FALSE)
}
\arguments{
\item{flag_debug}{Boolean - Enable debug statements}
}
\description{
profile_instrument_user_functions
Instrument user functions
}
