% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_is_instrumentation_enabled}
\alias{profile_is_instrumentation_enabled}
\title{profile_is_instrumentation_enabled
Return current instrumentation status}
\usage{
profile_is_instrumentation_enabled()
}
\value{
BOOLEAN - Instrumentation status
}
\description{
profile_is_instrumentation_enabled
Return current instrumentation status
}
