% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_print_eventlog}
\alias{profile_print_eventlog}
\title{profile_print_eventlog
Print global event log}
\usage{
profile_print_eventlog(df = pkg.env$PROFILE_EVENTLOG)
}
\arguments{
\item{df}{\itemize{
\item Dataframe object for eventlog
}}
}
\description{
profile_print_eventlog
Print global event log
}
