% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_print_instrumentation}
\alias{profile_print_instrumentation}
\title{profile_print_instrumentation
Print table of reduce dataframe}
\usage{
profile_print_instrumentation(flag_debug = TRUE)
}
\arguments{
\item{flag_debug}{Boolean - Enable debug header}
}
\description{
profile_print_instrumentation
Print table of reduce dataframe
}
