% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_replace_function}
\alias{profile_replace_function}
\title{profile_replace_function
Replace library function definition in package- and name-space}
\usage{
profile_replace_function(
  new_func,
  func_name,
  package_name,
  env_is_locked = TRUE
)
}
\arguments{
\item{new_func}{Function object with new definition}

\item{func_name}{String, name of function}

\item{package_name}{String, name of package function}

\item{env_is_locked}{Boolean - True if function name-/package- space is locked}
}
\description{
profile_replace_function
Replace library function definition in package- and name-space
}
