% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_reset_dataframe}
\alias{profile_reset_dataframe}
\title{profile_reset_dataframe
Reset function_count and function_time info to zero}
\usage{
profile_reset_dataframe(df)
}
\arguments{
\item{df}{Dataframe - Created by profile_create_dataframe() for collecting tracing info}
}
\description{
profile_reset_dataframe
Reset function_count and function_time info to zero
}
