% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_reset_eventlog}
\alias{profile_reset_eventlog}
\title{profile_reset_event_log
Assign new eventlog dataframe}
\usage{
profile_reset_eventlog()
}
\value{
Dataframe for eventlog
}
\description{
profile_reset_event_log
Assign new eventlog dataframe
}
