% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_save_dataframe}
\alias{profile_save_dataframe}
\title{profile_save_dataframe
Write dataframe to file (preferably reduced)}
\usage{
profile_save_dataframe(df, filename)
}
\arguments{
\item{df}{Dataframe - Created by profile_create_dataframe() for collecting tracing info}

\item{filename}{String - Filename to save output to}
}
\description{
profile_save_dataframe
Write dataframe to file (preferably reduced)
}
