% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{profile_save_eventlog}
\alias{profile_save_eventlog}
\title{profile_save_eventlog
Save event log to file}
\usage{
profile_save_eventlog(
  filename,
  df = pkg.env$PROFILE_EVENTLOG,
  flag_debug = FALSE
)
}
\arguments{
\item{filename}{String - Name of file to save eventlog to, .csv}

\item{df}{\itemize{
\item Dataframe object for eventlog
}}

\item{flag_debug}{Boolean - True to enable debug statements}
}
\description{
profile_save_eventlog
Save event log to file
}
