% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{profile_try_insert_instrumentation}
\alias{profile_try_insert_instrumentation}
\title{profile_try_insert_instrumentation
Checks function exceptions and calls profile_insert_instrumentation() if success}
\usage{
profile_try_insert_instrumentation(
  func_info,
  func_ptrs,
  env_is_locked,
  function_exception_list,
  function_methods_exception_list,
  flag_debug = F
)
}
\arguments{
\item{func_info}{Dataframe (struct) containing func_index info, func_name and packagE_name}

\item{func_ptrs}{Function[] - List of function objects}

\item{env_is_locked}{Boolean - True if function name-/package- space is locked}

\item{function_exception_list}{Object[] - List of function exceptions to skip}

\item{function_methods_exception_list}{Object[] - List of function method exceptions to skip}

\item{flag_debug}{Boolean - Enable debug output}
}
\description{
profile_try_insert_instrumentation
Checks function exceptions and calls profile_insert_instrumentation() if success
}
