% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quadprog_solveC}
\alias{quadprog_solveC}
\title{C++ function implemented by QuadProg++}
\usage{
quadprog_solveC(Dmat, dvec, At)
}
\arguments{
\item{Dmat, dvec, Amat}{See \code{\link[quadprog]{solve.QP}} for explanation.}
}
\value{
a list containing the solution
}
\description{
C++ function implemented by QuadProg++
}

