% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quadprog_solveR}
\alias{quadprog_solveR}
\title{C++ function to call the R function quadprog::solve.QP}
\usage{
quadprog_solveR(Dmat, dvec, Amat)
}
\arguments{
\item{Dmat, dvec, Amat}{See \code{\link[quadprog]{solve.QP}} for explanation.}
}
\value{
a list containing the solution
}
\description{
Call the R function quadprog::solve.QP to solve a quadratic programming problem
}

