% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.testthat.R
\name{quasi_capture}
\alias{quasi_capture}
\title{Method capturing the run}
\usage{
quasi_capture(quo, capture, label = NULL)
}
\arguments{
\item{quo}{an \code{rlang} quo}

\item{capture}{A function to derive the output / warnings / messages
of the function as e.g. \code{evaluate_promise}}

\item{label}{\code{character} A label for the evaluated value}
}
\value{
\code{act} A list including the label (\code{lab}),
  a caputre of the function (\code{cap}) and the code call
 itself as (\code{..})
}
\description{
Method capturing the run
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de})
}
