% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_pmpmeas_stop}
\alias{r_pmpmeas_stop}
\title{r_pmpmeas_stop}
\usage{
r_pmpmeas_stop(weight)
}
\arguments{
\item{weight}{Weight to scale metric values (eg average over N runs)}
}
\value{
R_NilValue
}
\description{
Wrapper for pmpmeas_stop
}
