% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_report_and_exit}
\alias{r_report_and_exit}
\title{r_report_and_exit}
\usage{
r_report_and_exit(msg)
}
\arguments{
\item{msg}{Error message to display}
}
\value{
R_NilValue
}
\description{
Wrapper for report_and_exit from R end
}
