% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{rbind}
\alias{rbind}
\title{Combine multiple PACdatasets by rows}
\usage{
rbind(...)
}
\arguments{
\item{...}{one or more PACdataset object.}
}
\value{
If params are PACdatasets with the same column definations, then will combine rows of PACds. See ?base::cbind for the value returned by the default methods.
}
\description{
Combine multiple PACdatasets by rows
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{subsetPACds}()}
}
\concept{PACdataset functions}
