% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{readBAM}
\alias{readBAM}
\title{Read the BAM file}
\usage{
readBAM(bam.path=NULL,bams=NULL,set.n=NULL,bam.keyword="*sorted.bam$",file=NULL,label=cell.type.name,group=rep(1,length(cell.type.name)))
}
\arguments{
\item{bam.path}{path of input of bam}

\item{bams}{bam files}

\item{set.n}{select the number of  bam files}

\item{bam.keyword}{selected bam files}
}
\value{
bam file
}
\description{
Read the BAM file
}
