% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{replace_user_function}
\alias{replace_user_function}
\title{replace_user_function}
\usage{
replace_user_function(new_func, func_name, package_name, env = .GlobalEnv)
}
\arguments{
\item{new_func}{Function, function object with new definition}

\item{func_name}{String, name of function}

\item{package_name}{String, name of package function}

\item{env}{Environment, environment function exists in (default .GlobalEnv)}
}
\description{
Replace user function definition
}
