% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{reset_eventlog}
\alias{reset_eventlog}
\title{reset_event_log
Assign new eventlog dataframe}
\usage{
reset_eventlog()
}
\value{
Dataframe for eventlog
}
\description{
reset_event_log
Assign new eventlog dataframe
}
