\name{rtest-package}
\alias{rtest-package}
\alias{rtest}
\docType{package}
\title{
\packageTitle{rtest}
}
\description{
\packageDescription{rtest}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{rtest}

Maintainer: \packageMaintainer{rtest}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = rtest)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{rtest}}
%\section{Documentation Index}{\packageIndices{rtest}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
