% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RTest.cat}
\alias{RTest.cat}
\title{Write Text To Console}
\usage{
RTest.cat(...)
}
\arguments{
\item{...}{Passed directly to \code{\link[base]{paste0}}.}
}
\description{
A message is written to the console if the option \code{RunitTestSuite_verbose} is set TRUE.
}
\seealso{
\code{\link{paste0}}, \code{\link[base]{cat}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
