% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RTest.print}
\alias{RTest.print}
\title{Print Text To Console}
\usage{
RTest.print(...)
}
\arguments{
\item{...}{Passed directly to \code{\link[base]{print}}.}
}
\description{
The arguments are printed if the option \code{RTest_verbose} is set TRUE.
}
\seealso{
\code{\link[base]{print}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
