% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtest.R
\docType{package}
\name{rtest}
\alias{rtest}
\alias{rtest-package}
\title{rtest: a test package}
\description{
The rtest package is a package I write to learn to write a package.
}
\section{rtest functions}{

The rtest package includes functions:
\link{hello}
\link{convolveR}
\link{convolveCpp}
}

