% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_evtWriter.R
\name{rTest_evtWriter_Write}
\alias{rTest_evtWriter_Write}
\title{rTest_evtWriter_Write - doxygen documentation
More info}
\usage{
rTest_evtWriter_Write(regionRef, event_type)
}
\arguments{
\item{regionRef}{Int - regionRef}

\item{event_type}{Bool - True if enter, False if exit}
}
\description{
rTest_evtWriter_Write - doxygen documentation
More info
}
