% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTest.R
\name{rTest_finalize}
\alias{rTest_finalize}
\title{rTest_finalize - doxygen documentation
Simple wrapper for C++ otf2 finalize function}
\usage{
rTest_finalize()
}
\description{
rTest_finalize - doxygen documentation
Simple wrapper for C++ otf2 finalize function
}
