% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTest_finalize_GlobalDefWriter}
\alias{rTest_finalize_GlobalDefWriter}
\title{finalize_GlobalDefWriter - doxygen documentation
More info}
\usage{
rTest_finalize_GlobalDefWriter(flag_debug = F)
}
\arguments{
\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
finalize_GlobalDefWriter - doxygen documentation
More info
}
