% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTest_globalDefWriter_WriteLocation}
\alias{rTest_globalDefWriter_WriteLocation}
\title{globalDefWriter_WriteLocation - doxygen documentation}
\usage{
rTest_globalDefWriter_WriteLocation(stringRef_name)
}
\arguments{
\item{stringRef_name}{Int - Index for name of location
More info}
}
\description{
globalDefWriter_WriteLocation - doxygen documentation
}
