% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTest_globalDefWriter_WriteRegion}
\alias{rTest_globalDefWriter_WriteRegion}
\title{globalDefWriter_WriteRegion - doxygen documentation
More info}
\usage{
rTest_globalDefWriter_WriteRegion(stringRef_RegionName, flag_debug = F)
}
\arguments{
\item{stringRef_RegionName}{Int - Index for names, taken from stringRef's}

\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\value{
regionRef Int - Index for regionRef
}
\description{
globalDefWriter_WriteRegion - doxygen documentation
More info
}
