% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTest_globalDefWriter_WriteRegions}
\alias{rTest_globalDefWriter_WriteRegions}
\title{globalDefWriter_WriteRegions - doxygen documentation
More info}
\usage{
rTest_globalDefWriter_WriteRegions(regionRef_array, regionRefName_array)
}
\arguments{
\item{regionRef_array}{Int[] - Index list for regionReg's}

\item{regionRefName_array}{Int[] - Index list for names, taken from stringRef's}
}
\description{
globalDefWriter_WriteRegions - doxygen documentation
More info
}
