% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTest_globalDefWriter_WriteString}
\alias{rTest_globalDefWriter_WriteString}
\title{globalDefWriter_WriteString - doxygen documentation
More info}
\usage{
rTest_globalDefWriter_WriteString(stringRefValue, flag_debug = F)
}
\arguments{
\item{stringRefValue}{String - String for globalDefWriter}

\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
globalDefWriter_WriteString - doxygen documentation
More info
}
