% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTest_globalDefWriter_WriteStrings}
\alias{rTest_globalDefWriter_WriteStrings}
\title{globalDefWriter_WriteStrings - doxygen documentation
More info}
\usage{
rTest_globalDefWriter_WriteStrings(stringRef_array, stringRefValue_array)
}
\arguments{
\item{stringRef_array}{Int[] - Index list for stringRef's}

\item{stringRefValue_array}{String[] - String list for globalDefWriter}
}
\description{
globalDefWriter_WriteStrings - doxygen documentation
More info
}
