% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloWorld.R
\name{rTest_helloWorld}
\alias{rTest_helloWorld}
\title{rTest_helloWorld - doxygen documentation
Simple wrapper for C++ hello world function}
\usage{
rTest_helloWorld()
}
\description{
rTest_helloWorld - doxygen documentation
Simple wrapper for C++ hello world function
}
