% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTest.R
\name{rTest_helloWorld_wrapper}
\alias{rTest_helloWorld_wrapper}
\title{rTest_helloWorld_wrapper - doxygen documentation}
\usage{
rTest_helloWorld_wrapper()
}
\description{
rTest_helloWorld_wrapper - doxygen documentation
}
