% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTest.R
\name{rTest_init}
\alias{rTest_init}
\title{rTest_init - doxygen documentation
Simple wrapper for C++ otf2 init function}
\usage{
rTest_init()
}
\description{
rTest_init - doxygen documentation
Simple wrapper for C++ otf2 init function
}
