% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_archive.R
\name{rTest_init_Archive}
\alias{rTest_init_Archive}
\title{rTest_init_Archive- doxygen documentation
Simple wrapper for C++ otf2 init function}
\usage{
rTest_init_Archive()
}
\description{
rTest_init_Archive- doxygen documentation
Simple wrapper for C++ otf2 init function
}
