% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTest.R
\name{rTest_init_GlobalDefWriter}
\alias{rTest_init_GlobalDefWriter}
\title{rTest_init_GlobalDefWriter - doxygen documentation
More info}
\usage{
rTest_init_GlobalDefWriter()
}
\description{
rTest_init_GlobalDefWriter - doxygen documentation
More info
}
