% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{class}
\name{RTestCollection-class}
\alias{RTestCollection-class}
\title{Class Definition 'RTestCollection'}
\description{
Class Definition 'RTestCollection'
}
\section{Slots}{

\describe{
\item{\code{project.name}}{(\code{character}) Name of the project.}

\item{\code{project.details}}{(\code{character}) Further details of the project.}

\item{\code{tester}}{(\code{character}) Name of the tester.}

\item{\code{test.start}}{(\code{character}) Start date of the testing project.}

\item{\code{collection}}{(\code{list}) The test case collection.}
}}

\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
