% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\name{RTestCollection}
\alias{RTestCollection}
\title{Constructor for 'RTestCollection'}
\usage{
RTestCollection(project.name = "RTest Execution",
  project.details = "Example test exectuion",
  project.tester = "Example tester")
}
\arguments{
\item{project.name}{- Name of the test campaign}

\item{project.details}{- Details of the test campaign}

\item{project.tester}{- Name of the tester}
}
\value{
\link{RTestCollection-class} object
}
\description{
Constructor for 'RTestCollection'
}
\examples{

obj <- RTestCollection()
show(obj)

}
\author{
Sebastian Wolf
}
