% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_evtWriter.R
\name{rTrace_evtWriter_Write}
\alias{rTrace_evtWriter_Write}
\title{rTrace_evtWriter_Write - doxygen documentation
More info}
\usage{
rTrace_evtWriter_Write(regionRef, event_type, flag_debug = F)
}
\arguments{
\item{regionRef}{Int - regionRef}

\item{event_type}{Bool - True if enter, False if exit}

\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
rTrace_evtWriter_Write - doxygen documentation
More info
}
