% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_archive.R
\name{rTrace_finalize_Archive}
\alias{rTrace_finalize_Archive}
\title{rTrace_finalize_Archive
Simple wrapper for C++ otf2 finalize function}
\usage{
rTrace_finalize_Archive(flag_debug = F)
}
\arguments{
\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
rTrace_finalize_Archive
Simple wrapper for C++ otf2 finalize function
}
