% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_evtWriter.R
\name{rTrace_finalize_EvtWriter}
\alias{rTrace_finalize_EvtWriter}
\title{rTrace_finalize_Archive- doxygen documentation
Simple wrapper for C++ otf2 init function}
\usage{
rTrace_finalize_EvtWriter(flag_debug = F)
}
\arguments{
\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
rTrace_finalize_Archive- doxygen documentation
Simple wrapper for C++ otf2 init function
}
