% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTrace_globalDefWriter_WriteLocation}
\alias{rTrace_globalDefWriter_WriteLocation}
\title{globalDefWriter_WriteLocation - doxygen documentation}
\usage{
rTrace_globalDefWriter_WriteLocation(stringRef_name, flag_debug = F)
}
\arguments{
\item{stringRef_name}{Int - Index for name of location
More info}

\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
globalDefWriter_WriteLocation - doxygen documentation
}
