% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTrace_globalDefWriter_WriteString}
\alias{rTrace_globalDefWriter_WriteString}
\title{globalDefWriter_WriteString - doxygen documentation
More info}
\usage{
rTrace_globalDefWriter_WriteString(stringRefValue, flag_debug = F)
}
\arguments{
\item{stringRefValue}{String - String for globalDefWriter}

\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
globalDefWriter_WriteString - doxygen documentation
More info
}
