% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloWorld.R
\name{rTrace_helloWorld}
\alias{rTrace_helloWorld}
\title{rTrace_helloWorld - doxygen documentation
Simple wrapper for C++ hello world function}
\usage{
rTrace_helloWorld()
}
\description{
rTrace_helloWorld - doxygen documentation
Simple wrapper for C++ hello world function
}
