% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_archive.R
\name{rTrace_init_Archive}
\alias{rTrace_init_Archive}
\title{rTrace_init_Archive
Simple wrapper for C++ otf2 init function}
\usage{
rTrace_init_Archive(flag_debug = F)
}
\arguments{
\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
rTrace_init_Archive
Simple wrapper for C++ otf2 init function
}
