% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTrace_init_GlobalDefWriter}
\alias{rTrace_init_GlobalDefWriter}
\title{init_GlobalDefWriter - doxygen documentation
More info}
\usage{
rTrace_init_GlobalDefWriter(flag_debug = F)
}
\arguments{
\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
init_GlobalDefWriter - doxygen documentation
More info
}
