% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{save_dataframe}
\alias{save_dataframe}
\title{save_dataframe
Write dataframe to file (preferably reduced)}
\usage{
save_dataframe(df, filename)
}
\arguments{
\item{df}{Dataframe - Created by create_dataframe() for collecting tracing info}

\item{filename}{String - Filename to save output to}
}
\description{
save_dataframe
Write dataframe to file (preferably reduced)
}
