% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_locationRef}
\alias{set_locationRef}
\title{set_locationRef}
\usage{
set_locationRef(id)
}
\arguments{
\item{id}{ID value belonging to this R proc}
}
\value{
R_NilValue
}
\description{
set_locationRef
}
