% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_maxUsedLocationRef_client}
\alias{set_maxUsedLocationRef_client}
\title{set_maxUsedLocationRef_client}
\usage{
set_maxUsedLocationRef_client(nprocs)
}
\arguments{
\item{nprocs}{Current number of active evtWriters/procs}
}
\value{
maxUsedLocationRef Current maximum evtWriters which were active so far
}
\description{
set_maxUsedLocationRef_client
}
