% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.R
\docType{methods}
\name{show,RTestCase-method}
\alias{show,RTestCase-method}
\title{Print Summary of the Test Case to Console}
\usage{
\S4method{show}{RTestCase}(object)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object.}
}
\description{
Print Summary of the Test Case to Console
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
