% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{skip_function}
\alias{skip_function}
\title{skip_function}
\usage{
skip_function(
  func_ptr,
  func_name,
  env,
  function_exception_list,
  function_methods_exception_list
)
}
\arguments{
\item{func_ptr}{Object - Pointer to function object}

\item{func_name}{String - Name of function}

\item{env}{Environment - Environment function is in}

\item{function_exception_list}{Object[] - List of functions to skip, generated by get_function_exception_list()}

\item{function_methods_exception_list}{Object[] - List of function methods to skip, generated by get_function_methods(function_exception_list)}
}
\value{
Boolean - TRUE if skip, else FALSE
}
\description{
Check if function should be skipped for instrumentation. Reasons
for skipping include function in FUNCTION_EXCEPTION_LIST, primitive
functions, functions used in instrumentation wrapper (to avoid recursion)
etc
}
