% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{subsetAPAMarkers}
\alias{subsetAPAMarkers}
\title{APA markers of individual cell types}
\usage{
subsetAPAMarkers(APAindex,celltype.anno=NULL,group.by,cluster.1=NULL,cluster.2=NULL,every.cluster=FALSE,genes=NULL,only.pos=FALSE,method= "wilcox",order="down")
}
\arguments{
\item{APAindex}{A data frame with rownames denoting genes, columns are samples, and values are the APA index values. Also includes other method-specific columns}

\item{celltype.anno}{A data frame denoting the cell classification information. Each row is one cell, and the column is the sample group
Row names in the celltype.anno need to match the column names of  APA index.}

\item{group.by}{Factor to group the cells by(the column name containing cell classification information, stored in celltype.anno)}

\item{cluster.1}{Vector of cell names belonging to group 1}

\item{cluster.2}{Vector of cell names belonging to group 2}

\item{every.cluster}{Finds APA markers for all cell types.When every.cluster=TRUE,cluster.1 and cluster.2 must be NULL}

\item{genes}{APA genes to test. Default is to use all APA genes}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{method}{Denotes which test to use. Available options are:
"wilcox" : Identifies differentially expressed APA genes between two groups of cells using a Wilcoxon Rank Sum test (default)
"t" : Identify differentially expressed APA genes between two groups of cells using the Student's t-test.
"LR" : Uses a logistic regression framework to determine differentially expressed APA genes.
 Constructs a logistic regression model predicting group membership based on each feature individually and compares this to a null model with a likelihood ratio test.}

\item{order}{When order=down,the returned values are sorted in descending order by avg_log2FC(The most significant marker in each group comes first);
When order=up,the returned values are sorted in ascending order by avg_log2FC(The most significant marker in each group ranks last)}
}
\value{
subsetAPAMarkers() reutrns a data frame containing APA markers information
}
\description{
Finds APA markers among different cell type
}
\examples{
#find all markers of ES
PAmakers.1<- subsetAPAMarkers(APAindex,celltype.anno=scPACds@colData ,group.by="celltype",cluster.1 = "ES", only.pos=FALSE,method= "wilcox")
#find all markers distinguishing ES from RS
PAmakers.2<- subsetAPAMarkers(APAindex,celltype.anno=scPACds@colData ,group.by="celltype",cluster.1 = "ES", cluster.2="RS",only.pos=FALSE,method= "wilcox")
#find markers for every cell cluster compared to all remaining cell clusters, report only the positive ones
APAmarkers<- subsetAPAMarkers(APAindex,celltype.anno=scPACds@colData ,group.by="celltype",every.cluster = TRUE, only.pos=TRUE,method= "wilcox")
}
