% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sugerTest}
\alias{sugerTest}
\title{Function to test the Rcpp suger}
\usage{
sugerTest(N, mu, sigma)
}
\arguments{
\item{N}{number of samples}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
Gaussian samples
}
\description{
Function to test the Rcpp suger
}

