% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{summary,RTestCollection-method}
\alias{summary,RTestCollection-method}
\title{Execution Summary As R Data Object}
\usage{
\S4method{summary}{RTestCollection}(object, test.TCs = NULL)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object.}

\item{test.TCs}{(\code{character}) Vector with the TCs to be executed or NULL if all
all TCs of the collection should be tested.}
}
\description{
Execution Summary As R Data Object
}
\seealso{
\code{\link{RTestCollection-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
