% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{systemInfo.host}
\alias{systemInfo.host}
\title{Summarize Host System Information}
\usage{
systemInfo.host()
}
\value{
(\code{data.frame}) A character vector containing the R version information.
}
\description{
This method creates a tabular listing of current host system.
}
\seealso{
\code{\link[base]{Sys.info}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
