% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{systemInfo.packages}
\alias{systemInfo.packages}
\title{Summarize System Packages}
\usage{
systemInfo.packages(which = "loadedOnly")
}
\arguments{
\item{which}{(\code{character}) Specifies, which packages to display. One of 
'\code{basePkgs}', '\code{loadedOnly}' or '\code{otherPkgs}' (for 
details see \code{\link[utils]{sessionInfo}}.}
}
\value{
(\code{data.frame}) A table containing the packages' names, versions and build dates.
}
\description{
This method creates a tabular listing of the packages, which are currently loaded and available
to the system
}
\seealso{
\code{\link[utils]{sessionInfo}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
