% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{systemInfo.RInst}
\alias{systemInfo.RInst}
\title{Summarize R Version Information}
\usage{
systemInfo.RInst()
}
\value{
(\code{character}) A character vector containing the R version information.
}
\description{
This method creates a tabular listing of current R version information.
}
\seealso{
\code{\link[utils]{sessionInfo}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
