% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.test-adapter.R
\name{test.RTest.funct_01}
\alias{test.RTest.funct_01}
\alias{test.RTest.funct_01,RTestCase-method}
\title{Test Function For Testing Function 'RTest::test_returnValue_data.frame_cellbycell'}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCase-class}} object}
}
\value{
(\code{list})
}
\description{
Test Function For Testing Function 'RTest::test_returnValue_data.frame_cellbycell'
}
\seealso{
\code{\link{RTestCase-class}}
}
\author{
Sebastian Wolf
}
