% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test__ports}
\alias{test__ports}
\title{Test ports used by rTrace
Intended for developers}
\usage{
test__ports(
  port_type,
  ports = as.numeric(c(5554, 5556, 5557, 5559)),
  num_messages = 1000L,
  message_size = 8L
)
}
\arguments{
\item{port_type}{1:=req-rep, 2:=push-pull, 3:=pub-sub, else error}

\item{ports}{List of ports to test on valid range [1024,49151], default {5554,5556,5557,5559}}

\item{num_messages}{Number of messages to send}

\item{message_size}{Size of each msg in KB}
}
\value{
0 if success, else non-zero
}
\description{
Test ports used by rTrace
Intended for developers
}
