% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test__signal}
\alias{test__signal}
\title{test__signal - Function to test handling of SIGRTRACE thrown by report_and_exit()}
\usage{
test__signal(signal_proc, is_master = TRUE)
}
\arguments{
\item{signal_proc}{0 if throwing signal from master proc, 1 from server, else from slave}

\item{is_master}{True if run from master R proc, else false if slave (default: true for single-threaded cases)}
}
\value{
int 0 if success
}
\description{
test__signal - Function to test handling of SIGRTRACE thrown by report_and_exit()
}
