% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test__struct_size}
\alias{test__struct_size}
\title{Verify structs have unique sizes in order to distinguish between events or errors.
Intended for developers}
\usage{
test__struct_size()
}
\value{
0 if success, else non-zero
}
\description{
Verify structs have unique sizes in order to distinguish between events or errors.
Intended for developers
}
