% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCase.test-adapter.R
\name{test_fun}
\alias{test_fun}
\title{A simple function to Test the RTest package}
\usage{
test_fun(dat, mult)
}
\arguments{
\item{dat}{(\code{data.frame}) Any simple number dataframe with minimum one column}

\item{mult}{(\code{numeric}) Any simple number}
}
\value{
A Table with the number vector + a sum of the vector multiplied by \code{mult}
}
\description{
A simple function to Test the RTest package
}
\examples{

dat <- data.frame(x=c(1,1))
mult <- 1
test_fun(dat,mult)

}
\author{
Sebastian Wolf
}
