% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{test_instrumentation}
\alias{test_instrumentation}
\title{test_instrumentation}
\usage{
test_instrumentation(func_ptr, func_name, expr, flag_debug = F)
}
\arguments{
\item{func_ptr}{Object - function to test}

\item{func_name}{String - name of function}

\item{expr}{Expression - R expression to call to test func}

\item{flag_debug}{Boolean - Enable debug statements}
}
\description{
Utility function - Call function with expr and output dataframe changes if enabled
}
