% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_it}
\alias{test_it}
\title{Make ATAC CDS object}
\usage{
test_it()
}
\description{
This function takes as input a data frame or a path to a file in a sparse
matrix format and returns a properly formatted \code{CellDataSet} (CDS)
object.
}
