% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.test.R
\name{test_manualCheck_confirmWindow}
\alias{test_manualCheck_confirmWindow}
\title{Creates a Tcl/Tk Window for confirmation of the manual check}
\usage{
test_manualCheck_confirmWindow(openrecexp = NULL, expectedTxt = NULL)
}
\arguments{
\item{openrecexp}{(\code{function}) Function to open a window with the received and expected
values or NULL to hide all.}

\item{expectedTxt}{(\code{character}) Shown as text for the expected vidow.}
}
\value{
(\code{list}) List with elements 'result': TRUE or FALSE depending on the user's
          choice and 'comment': inserted user comment.
}
\description{
Creates a Tcl/Tk Window for confirmation of the manual check
}
\section{Package dependencies}{

The \code{tcltk} package is required. Please install it from CRAN if not in your current library
path.
}

\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
