% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.test.R
\name{test_manualCheck_file}
\alias{test_manualCheck_file}
\title{Creates a manual check via tcl/tk interface of created output file}
\usage{
test_manualCheck_file(result, reference, xmlTestSpec = NULL,
  add.desc = NULL, openrecexp = NULL, openrecexp.exec = FALSE)
}
\arguments{
\item{result}{(\code{object}) The result object to be tested.}

\item{reference}{(\code{object}) The reference object.}

\item{xmlTestSpec}{(\code{XMLNode}) The XML definition of type 'RTestTest_variable'.}

\item{add.desc}{(\code{character}) Additional description added to the XML
definition.}

\item{openrecexp}{(\code{code}) Function to open/show reference and expected result.}

\item{openrecexp.exec}{(\code{logical}) Specify, if the openrecexp-function should be
executed by default or only on user decision (by click).}
}
\description{
Creates a manual check via tcl/tk interface of created output file
}
\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
