% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test__ports.R
\name{test_ports}
\alias{test_ports}
\title{test_ports}
\usage{
test_ports(
  port_type,
  ports = c(5554, 5555, 5556, 5557, 5559),
  num_messages = 1000,
  message_size = 8
)
}
\arguments{
\item{port_type}{1:=req-rep, 2:=push-pull, 3:=pub-sub, else error}

\item{ports}{List of ports to test on valid range [1024,49151], default {5554,5555,5556,5557,5559}}

\item{num_messages}{Number of messages to send}

\item{message_size}{Size of each msg in KB}
}
\description{
test_ports
}
