% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.test.R
\name{test_returnValue_any}
\alias{test_returnValue_any}
\title{Generically compare two values with RTest}
\usage{
test_returnValue_any(result, reference, xmlTestSpec)
}
\arguments{
\item{result}{(\code{any}) Any value of type character, numeric, data.frame or list
(image links do not work!)}

\item{reference}{(\code{any}) Any value of type character, numeric, data.frame or list
(image links do not work!)}

\item{xmlTestSpec}{(\code{XMLNode}) An XMLNode of type \code{RTest_test_returnValue_...}}
}
\value{
The function will not return anything but call \code{testthat} functions
   creating outputs in the reporter
}
\description{
This function compares two value by a \code{test_returnValue_...} function
that fits the class of the \code{reference} input parameter.
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
