% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.test.R
\name{test_returnValue_data.frame_cellbycell}
\alias{test_returnValue_data.frame_cellbycell}
\title{Tests a Standard R 'data.frame' Cell-By-Cell ('RTestTest_data.frame_cellbycell')}
\usage{
test_returnValue_data.frame_cellbycell(result, reference, xmlTestSpec,
  add.desc = NULL)
}
\arguments{
\item{result}{(\code{data.frame}) The result data.frame to be tested}

\item{reference}{(\code{data.frame}) The reference data.frame}

\item{xmlTestSpec}{(\code{XMLNode}) The XML definition of type
'RTestTest_data.frame_cellbycell'}

\item{add.desc}{(\code{character}) Additional description added to the XML
definition.}
}
\description{
Tests a Standard R 'data.frame' Cell-By-Cell ('RTestTest_data.frame_cellbycell')
}
\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
