% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.test.R
\name{test_returnValue_list_nodebynode}
\alias{test_returnValue_list_nodebynode}
\title{Tests a Standard R 'list' Node-By-Node ('RTestTest_list_nodebynode')}
\usage{
test_returnValue_list_nodebynode(result, reference, xmlTestSpec,
  add.desc = NULL)
}
\arguments{
\item{result}{(\code{list}) The result list to be tested}

\item{reference}{(\code{list}) The reference list}

\item{xmlTestSpec}{(\code{XMLNode}) The XML definition of type
'RTestTest_list_nodebynode'}

\item{add.desc}{(\code{character}) Additional description added to the XML
definition.}
}
\description{
Tests a Standard R 'list' Node-By-Node ('RTestTest_list_nodebynode')
}
\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Sergej Potapov \email{sergej.potapov@roche.com}
}
