% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{Function timesTwo}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the vector x*2
}
\description{
the function to multipy x by 2
}
\examples{
timesTwo(10)
}

