% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{total_num_functions}
\alias{total_num_functions}
\title{total_num_functions
Find total number of loaded packages and functions}
\usage{
total_num_functions(debug_flag = FALSE, flag_user_functions = FALSE)
}
\arguments{
\item{debug_flag}{Boolean - TRUE to print debug information}

\item{flag_user_functions}{Boolean - True if also flagging user functions}
}
\value{
Int[] Number of functions per package
}
\description{
total_num_functions
Find total number of loaded packages and functions
}
