% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transpose}
\alias{transpose}
\title{test for matrix transpose}
\usage{
transpose(Amat)
}
\arguments{
\item{Amat}{a matrix to be transposed}
}
\value{
the transpose of matrix Amat
}
\description{
test for matrix transpose
}

