% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{useGff}
\alias{useGff}
\title{Parse genome annotation file}
\usage{
useGff(gff, txdb,species)
}
\arguments{
\item{gff}{annotation gtf/gff3/gff fotmat  file}

\item{txdb}{a TxDb object}

\item{species}{specify species, then specify the  source of gff file from BioMart.}
}
\value{
If input gtf.file,then return parsed genome annotation object, which is a data.frame
If input txdb ,then return parsed genome annotation object, which is a list of three elements (anno.rna, anno.need, anno.frame).
If input a species, then return gff file from BioMart.
}
\description{
parse genome annotation file of  gtf/gff3/gff format or genome annotation object of TxDb.
 If specify species, then specify the  source of gff file from BioMart.
}
\examples{
## Way1: Based on an annotation file in gtf format
gff.path <- "Arabidopsis_thaliana.TAIR10.49.gtf"
gtf.data <- useGff(gff =gff.path)

##Way2: Based on a TxDb object generated from BioMart.
# Parse Arabidopsis Txdb
library(TxDb.Athaliana.BioMart.plantsmart28)
txdb <- useGff(TxDb.Athaliana.BioMart.plantsmart28)
# Parse mm10 Txdb
BiocManager::install("TxDb.Mmusculus.UCSC.mm10.ensGene")
library(TxDb.Mmusculus.UCSC.mm10.ensGene)
txdb <- useGff(TxDb.Mmusculus.UCSC.mm10.ensGene)
}
