% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{vizAPAMarkers}
\alias{vizAPAMarkers}
\title{Visualize APA markers in individual cell types.}
\usage{
vizAPAMarkers(APAindex,celltype.anno,group.by,markers=NULL,type="heatmap",vln.flip=TRUE,vln.fill.by="celltype",vln.cols=NULL,dot.cols=c("lightgrey", "#9b393b"),hm.group.cols=NULL)
}
\arguments{
\item{APAindex}{A data frame with rownames denoting genes, columns are samples, and values are the APA index values. Also includes other method-specific columns}

\item{celltype.anno}{A data frame denoting the cell classification information. Each row is one cell, and the column is the sample group
Row names in the celltype.anno need to match the column names of  APA index.}

\item{group.by}{Factor to group the cells by(the column name containing cell classification information, stored in celltype.anno)}

\item{markers}{APA markers to plot}

\item{type}{violin/heatmap/dot:specified drawing type.When type=violin/heatmap/dot,show APA markers in the form of violin plot/heatmap/bubble plot}

\item{vln.flip}{flip plot orientation.If type=violin and vln.flip=TRUE,cell classes on the x-axis}

\item{vln.fill.by}{celltype/markers:If type=violin,color violins based on either 'celltype' or 'markers'}

\item{vln.cols}{If type=violin,colors to use for plotting}

\item{dot.cols}{If type=dot,colors to use for plotting:the two colors to form the gradient over}

\item{hm.group.cols}{If type=heatmap,colors to use for the color bar:add a color bar showing group status for cells}
}
\value{
vizAPAMarkers() reutrns violin plot/heatmap/bubble plot
}
\description{
Draws a violin point/heatmap/bubble  plot to show APA markers
}
\examples{
#specified drawing type
topAPAM<-APAmarkers \%>\% group_by(cluster) \%>\% top_n(n =2, wt = avg_log2FC)
vizAPAMarkers(APAindex,celltype.anno=scPACds@colData ,group.by="celltype",type="heatmap",markers=topAPAM$gene)
#If type=violin,color violins based on 'celltype'
vizAPAMarkers(APAindex,celltype.anno=scPACds@colData ,group.by="celltype",type="violin",markers=topAPAM$gene)
#If type=violin,color violins based on 'markers'
vizAPAMarkers(APAindex,celltype.anno=scPACds@colData ,group.by="celltype",type="violin",markers=topAPAM$gene,vln.fill.by="markers")
}
