% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{vizPAplot}
\alias{vizPAplot}
\alias{vizBubble}
\title{box/violin/point plot}
\usage{
vizPAplot(scPACds,datatype,group,object="chr",type="point",log=TRUE,cols=NULL)

vizBubble(PAdata,datatype,genes,celltype.anno=NULL,group.by,conds=NULL,col=c("lightgrey", "#9b393b"))
}
\arguments{
\item{scPACds}{A dataset of the PACdataset class}

\item{datatype}{PA/APA.When datatype=PA,the PA data provided is PA count;When datatype=APA,the PA data provided is APA index}

\item{group}{the column name containing cell classification information, stored in colData}

\item{object}{chr/gene.Specify whether the drawing object is chromosome or genetic PA sites}

\item{type}{box/violin/point:specified drawing type .If type=point and object=chr,show it in the form of box plot, violin plot plus point plot;
if  type=point and object=gene, show it with point plot}

\item{log}{Plot the y axis value on log scale(calculate the natural logarithm of the raw data)}

\item{cols}{Colors to use for plotting.When cols=NULL,use the default color matching}

\item{PAdata}{A dataset of the PACdataset class(eg:scPACds) or APA index(uses all samples to filter APA sites, and then gets index for each sample):should be a data frame}

\item{genes}{Select the genes}

\item{group.by}{Factor to group the cells by(the column name containing cell classification information, stored in PACdataset@colData or celltype.anno)}

\item{celltype.anno}{A data frame denoting the cell classification information. Each row is one cell, and the column is the sample group
Row names in the celltype.anno need to match the column names of the PACdataset@counts or APA index.
when PAdata=PACdataset and colDataFile is provided(the information is stored in the colData at this time), celltype.anno=NULL}

\item{conds}{Select the type of cells needs to be plotted}

\item{col}{Colors to plot:the name of a palette from RColorBrewer::brewer.pal.info or a pair of colors defining a gradient}
}
\value{
vizPAplot() reutrns box/violin/point plot

vizBubble() reutrns bubble plot
}
\description{
Draws a box/violin/point plot to show all PA expression data(PA count/PA ratio) in a chromosomal region or gene across different cell types

Draws a bubble plot of single cell data (PA count,APA index)
}
\examples{
#way1.plot the chromosomal region
allPASC<-subsetPACds(scPACds,group = "celltype",chrs="chr12",noIntergenic=TRUE,pool=TRUE)
vizPAplot(scPACds =allPASC,datatype="count",group="celltype",object="chr",type="point",log=TRUE)
#way2.plot the gene
PAgene<-subsetPACds(scPACds,group = "celltype",noIntergenic=TRUE,genes="ENSMUSG00000021180")
vizPAplot(scPACds =PAgene,datatype="count",group="celltype",object="gene",type="box")
#show  PA expression data(PA count) in the gene across different cell types
vizBubble(PAdata=scPACds,datatype="PA",genes="ENSMUSG00000019969",group.by="celltype")
#change the color:the color name from RColorBrewer::brewer.pal.info
vizBubble(PAdata=scPACds,datatype="PA",genes="ENSMUSG00000019969",group.by="celltype",col="Set2")
#select the cell type
vizBubble(PAdata=scPACds,datatype="PA",genes="ENSMUSG00000019969",group.by="celltype",conds=c("ES","RS"),col=c("lightgrey", "green"))
}
