% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{vizTracks}
\alias{vizTracks}
\title{track plot}
\usage{
vizTracks(gtf.data,bams,pas.postition,region,gene,PA, space5=300,space3=300,
                 showGene=TRUE, showallPA=FALSE,bm=bm,Group=FALSE,genome.version=NULL,
                 cov.Merge="No",showCov="Single",covStyle="cov",
                 bams.col="common",bams.fill="common",bams.background.col="common",
                 PAcol="red",PAfill="red",
                 cov.Merge.background.col="red",cov.Merge.col="red",cov.Merge.fill="red")
}
\arguments{
\item{gtf.data}{annotation gtf fotmat  file}

\item{bams}{bam file,it is used to extract the expression level of PAS to plot the PAS coverage information}

\item{pas.position}{poly(A) site information}

\item{region}{specify the chromosomal region}

\item{gene}{specify the gene}

\item{space5}{extend 5' end,default is 300 bp}

\item{space3}{extend 3' end,default is 300 bp}

\item{showGene}{If TRUE,tnen the area where the map is drawn becomes the gene area where the PA is located
If FALSE,then only show that part of the area where the PA is located}

\item{showallPA}{If TRUE,then  show all PAs.Otherwise Only show the PA specified under this gene}

\item{Group}{default is FALSE.If TRUE,then plot coverage models on one track.If FALSE, then each condition is one track.}

\item{bm}{BioMart database}

\item{genome.version}{genome.version}

\item{cov.Merge}{Avg/Sum/No.If cov.Merge=Avg,then  average the coverage models on one track .
If cov.Merge=Sum,then sum the coverage models on one track.If cov.Merge=NO,then make a separate track.}

\item{showCov}{Both/Single/Merge.default is Single.If showCov=Both and cov.Merge=Avg/Sum,then both the merged  and the separate coverage model are showed.
If showCov=Single,then only show the separate coverage model.
If showCov=Merge and cov.Merge=Avg/Sum,then only show the merged coverage model.}

\item{covStyle}{cov/curve.If covStyle=cov,then  the coverage model draws in the form of an overlay.
If covStyle=curve ,the coverage model draws in the form of a curve.}

\item{pA}{specify the poly (A) site}
}
\value{
vizTracks() reutrns Gviz's Tracks.If gtf.data is provided,then plot the gene model.
If pas.position is provided,then piot the pA model.
If bam file is provided,then plot the coverage model across different cell types.
}
\description{
show the gene model of a gene or a chromosomal region,the PAS model and the coverage  model of a specified PAS across different cell types
}
\examples{
#way1.specify the chromosomal region
vizTracks(gtf.data=gtf.data,bams=bams,pas.postition=pas.postition,region="chr1:+:3000:9000", bm=bm，cov.Merge = "Avg",showCov = "Both")
#way2.specify the gene
vizTracks(gtf.data = gtf.data,pas.position = pas.postition,bams=bams,gene="AT1G01040",bm=bm,cov.Merge = "Avg",showCov = "Both")
#way3.specify the poly (A) site
vizTracks(gtf.data=gtf.data,bams=bams,pas.position = pas.position,PA="chr1:+:10000447",bm=bm,
          genome.version = genome.version,showGene = FALSE,showallPA = TRUE,cov.Merge = "Avg",showCov = "Both")
}
