% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtest_function.R
\name{vizUMAP}
\alias{vizUMAP}
\title{UMAP plot}
\usage{
vizUMAP(PAdata,celltype.anno=NULL,group.by,PA = NULL,gene=NULL,dims=1:10,order=NULL,cols=NULL,order.cols=NULL,pt.cols= c("lightgrey", "#185827"),size=NULL)
}
\arguments{
\item{PAdata}{A dataset of the PACdataset class(eg:scPACds) or APA index(uses all samples to filter APA sites, and then gets index for each sample):should be a data frame}

\item{celltype.anno}{A data frame denoting the cell classification information. Each row is one cell, and the column is the sample group
Row names in the celltype.anno need to match the column names of the PACdataset@counts or APA index.
when PAdata=PACdataset and colDataFile is provided(the information is stored in the colData at this time), celltype.anno=NULL}

\item{group.by}{Factor to group the cells by(the column name containing cell classification information, stored in PACdataset@colData or celltype.anno)}

\item{PA}{Select the poly(A) site to show it's expression in UMAP Plot}

\item{gene}{Select the APA gene to show the usage of PACs in UMAP Plot}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{order}{Select the cell types that highlight in UMAP Plot}

\item{cols}{Vector of colors, each color corresponds to an identity class}

\item{order.cols}{The colors of the cell types that highlight(if order is not NULL)}

\item{pt.cols}{When PA or gene is not NULL ,the color of the point in UMAP Plot:the two colors to form the gradient over}

\item{size}{Adjust point size for plotting}
}
\value{
vizUMAP() reutrns UMAP plot
}
\description{
Graphs the output of a dimensional reduction technique on a 2D scatter plot where each point is a cell and it's positioned based on the cell embeddings determined by the reduction technique.
By default, cells are colored by their celltype (can be changed with the group.by parameter)
}
\examples{
vizUMAP(scPACds,group.by = "celltype",dims=1:5)
#Select the cell types that highlight in UMAP Plot
vizUMAP(scPACds,group.by = "celltype",dims=1:5,order=c("ES","RS"))
#Select the poly(A) site to show it's expression in UMAP Plot
vizUMAP(scPACds,group.by = "celltype",dims=1:5,PA="PA3346")
}
