% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTestCollection.R
\docType{methods}
\name{writeExecSummary.html}
\alias{writeExecSummary.html}
\alias{writeExecSummary.html,RTestCollection-method}
\title{Write Summary of Last Test Case Executions as HTML}
\usage{
\S4method{writeExecSummary.html}{RTestCollection}(object, out.fPath,
  test.TCs = NULL, open = TRUE, report.onlyFailed = FALSE,
  logo = NULL)
}
\arguments{
\item{object}{(\code{object}) The \code{\link{RTestCollection-class}} object.}

\item{out.fPath}{(\code{character}) Path to output file.}

\item{test.TCs}{(\code{character}) Vector with the TCs to be executed or NULL if all
all TCs of the collection should be tested.}

\item{open}{(\code{logical}) Should the generated file be opened (TRUE) or not
(FALSE) after report generation.}

\item{report.onlyFailed}{(\code{logical}) Report only failed exceptions (TRUE) or
all exceptions (FALSE, default).}

\item{logo}{(\code{character}) Path to alternative logo file. To use the default
logo, use NULL. To use no logo use NA.}
}
\description{
Write Summary of Last Test Case Executions as HTML
}
\seealso{
\code{\link{RTestCollection-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
