% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlRead.R
\name{xmlReadData_list}
\alias{xmlReadData_list}
\title{Read XML Data From Type 'xmlReadData_list' as R 'list' (DUMMY)}
\usage{
xmlReadData_list(xmlItem)
}
\arguments{
\item{xmlItem}{(\code{XMLNode}) Object of class \code{XMLNode} that defines the 
a list object and fullfills XSD definition 'xmlReadData_list'.}
}
\value{
(\code{data.frame})
}
\description{
DUMMY: IMPLEMENTATION HAS TO BE DONE!!
}
\examples{
data <- '<mylist>

<data.frame name="ITEM1">
<col-defs>
  <coldef name="Column1" type="character"/>
  <coldef name="Column2" type="numeric"/>
</col-defs>
<row name="1"><cell>ID1</cell><cell>1</cell></row>
<row name="2"><cell>ID2</cell><cell>2.1</cell></row>
<row name="3"><cell>ID3</cell><cell>3.1</cell></row>
</data.frame>
<vector name="ITEM2" type="numeric"><element>1</element><element>2</element></vector>

</mylist>
'
item <- XML::xmlRoot(XML::xmlParse(data,asText=TRUE))
value <- RTest::xmlReadData_list(item)
stopifnot(names(value)[1]=="ITEM1")
stopifnot(length(value[[2]])==2)
}
\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
