% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlRead.R
\name{xmlReadData_text}
\alias{xmlReadData_text}
\title{Read XML Data From Type 'xmlReadData_text' as R Variable}
\usage{
xmlReadData_text(xmlItem)
}
\arguments{
\item{xmlItem}{(\code{XMLNode}) Object of class \code{XMLNode} that defines the 
a simple variable and fullfills XSD definition 'xmlReadData_text'.}
}
\value{
(\code{vector})
}
\description{
Read XML Data From Type 'xmlReadData_text' as R Variable
}
\examples{
data <- '<variable type="character">My text is awesome</variable>'
item <- XML::xmlRoot(XML::xmlParse(data,asText=TRUE))
value <- RTest::xmlReadData_text(item)
stopifnot(value=="My text is awesome")
}
\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
