% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlRead.R
\name{xmlReadData_variable}
\alias{xmlReadData_variable}
\title{Read XML Data From Type 'xmlReadData_variable' as R Variable}
\usage{
xmlReadData_variable(xmlItem)
}
\arguments{
\item{xmlItem}{(\code{XMLNode}) Object of class \code{XMLNode} that defines the 
a simple variable and fullfills XSD definition 'xmlReadData_variable'.}
}
\value{
(\code{vector})
}
\description{
Read XML Data From Type 'xmlReadData_variable' as R Variable
}
\examples{
data <- '<variable name="myvar" value="4" type="numeric"/>'
item <- XML::xmlRoot(XML::xmlParse(data,asText=TRUE))
value <- RTest::xmlReadData_variable(item)
print("5 shall be the outcome")
print(value + 1)
}
\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
