% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlRead.R
\name{xmlReadData_vector}
\alias{xmlReadData_vector}
\title{Read XML Data From Type 'xmlReadData_vector' as R Vector}
\usage{
xmlReadData_vector(xmlItem)
}
\arguments{
\item{xmlItem}{(\code{XMLNode}) Object of class \code{XMLNode} that defines the 
a vector and fullfills XSD definition 'xmlReadData_vector'.}
}
\value{
(\code{vector})
}
\description{
Read XML Data From Type 'xmlReadData_vector' as R Vector
}
\examples{
data <- '<testvector type="numeric"><element>1</element><element>2</element></testvector>'
item <- XML::xmlRoot(XML::xmlParse(data,asText=TRUE))
value <- RTest::xmlReadData_vector(item)
print("2 shall be the outcome")
print(length(value))
}
\seealso{
\code{\link[XML]{XMLNode-class}}
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
