% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteContext}
\alias{xmlWriteContext}
\title{Write the Opening (Header, Root-Tag) and Closing for a RTestCase}
\usage{
xmlWriteContext(TCType, id, opening = TRUE, closing = TRUE,
  xsd.scheme = NULL, printXML = TRUE)
}
\arguments{
\item{TCType}{(\code{character}) TC Type}

\item{id}{(\code{character}) TC ID}

\item{opening, closing}{(\code{logical}) Specify if the opening and/or the closing
tags should be written.}

\item{xsd.scheme}{(\code{character}) Path to XSD Scheme}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{list}) Opening [[1]] and Closing [[2]] of the Test Case
}
\description{
Write the Opening (Header, Root-Tag) and Closing for a RTestCase
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
