% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteData_data.frame}
\alias{xmlWriteData_data.frame}
\title{Write a R 'data.frame' as XML Data of Type 'xmlReadData_data.frame'}
\usage{
xmlWriteData_data.frame(elemname = "data", data, name = NULL,
  printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The name of the element's root tag}

\item{data}{(\code{data.frame}) The data to write}

\item{name}{(\code{character}) The data name.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write a R 'data.frame' as XML Data of Type 'xmlReadData_data.frame'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
