% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteData_list}
\alias{xmlWriteData_list}
\title{Write a R 'list' as XML Data of Type 'xmlReadData_list'}
\usage{
xmlWriteData_list(elemname = "list", data, name = NULL,
  printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The name of the element's root tag}

\item{data}{(\code{ANY}) The list to write}

\item{name}{(\code{character}) The name of the list}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write a R 'list' as XML Data of Type 'xmlReadData_list'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
