% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteData_matrix}
\alias{xmlWriteData_matrix}
\title{Write a R 'matrix' as XML Data of Type 'xmlReadData_matrix'}
\usage{
xmlWriteData_matrix(elemname = "data", data, name = NULL,
  printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The name of the element's root tag}

\item{data}{(\code{matrix}) The data to write}

\item{name}{(\code{character}) The data name.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write a R 'matrix' as XML Data of Type 'xmlReadData_matrix'
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
