% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteData_variable}
\alias{xmlWriteData_variable}
\title{Write a R 'constat' as XML Data of Type 'xmlReadData_variable'}
\usage{
xmlWriteData_variable(elemname = "variable", data, name = NULL,
  printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The name of the element's root tag}

\item{data}{(\code{ANY}) The variable to write}

\item{name}{(\code{character}) The name of the variable}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write a R 'constat' as XML Data of Type 'xmlReadData_variable'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
