% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteData_vector}
\alias{xmlWriteData_vector}
\title{Write a R 'vector' as XML Data of Type 'xmlReadData_vector'}
\usage{
xmlWriteData_vector(elemname = "vector", data, name = NULL,
  printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The name of the element's root tag}

\item{data}{(\code{vector}) The vector data to write}

\item{name}{(\code{character}) The data name.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write a R 'vector' as XML Data of Type 'xmlReadData_vector'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
