% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteInputData}
\alias{xmlWriteInputData}
\title{Write the Input-Data section for a RTestCase}
\usage{
xmlWriteInputData(..., printXML = TRUE)
}
\arguments{
\item{...}{(\code{character}) Stuff to include in the input section}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write the Input-Data section for a RTestCase
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
