% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteSynopsis}
\alias{xmlWriteSynopsis}
\title{Write the Synopsis for a RTestCase}
\usage{
xmlWriteSynopsis(version, author, shortDescription = NULL,
  description = NULL, creationDate = NULL, changes = list(list(author
  = author, date = creationDate, desc = "Initial Version")),
  label = NULL, printXML = TRUE)
}
\arguments{
\item{version}{(\code{dcharacter}) Version Number}

\item{author}{(\code{character}) Author}

\item{shortDescription}{(\code{character}) Short description}

\item{description}{(\code{character}) Description}

\item{creationDate}{(\code{character}) Creation Date}

\item{changes}{(\code{list}) A listing of lists each representing one change with
attributes 'author', 'date' and 'desc'}

\item{label}{(\code{character}) Labels.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write the Synopsis for a RTestCase
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
