% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTest}
\alias{xmlWriteTest}
\title{Write the Test section for a RTestCase}
\usage{
xmlWriteTest(elemname, testdesc = NA, ..., printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The name of the element's root tag}

\item{testdesc}{(\code{character}) The description of the test's root tag}

\item{...}{(\code{character}) Stuff to include in the test section}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write the Test section for a RTestCase
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
