% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTest_data.frame_cellbycell}
\alias{xmlWriteTest_data.frame_cellbycell}
\title{Write XML Test Definition of Type 'RTestTest_data.frame_cellbycell'}
\usage{
xmlWriteTest_data.frame_cellbycell(elemname = "test",
  desc = "testname", data, diff_type = "absolute", tolerance = 0,
  compare_type = "equal", printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The tag name of the test}

\item{desc}{(\code{character}) Description}

\item{data}{(\code{data.frame}) The reference data for which the test should be 
written.}

\item{diff_type}{(\code{character}) Difference 'absolute' or 'relative' that is used for comparison.}

\item{tolerance}{(\code{ANY}) Named vector with tolerances, single entry if same for all.}

\item{compare_type}{(\code{character}) Comparator used in the XML spec.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write XML Test Definition of Type 'RTestTest_data.frame_cellbycell'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
