% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTest_execution}
\alias{xmlWriteTest_execution}
\title{Write XML Test Definition of Type 'RTestTest_execution'}
\usage{
xmlWriteTest_execution(elemname = "execution", desc = NULL,
  executionType = "silent", printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The tag name of the test}

\item{desc}{(\code{character}) The testname}

\item{executionType}{(\code{character}) The execution mode to be checked (i.e. 'silent', 
'warning', 'error').}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write XML Test Definition of Type 'RTestTest_execution'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
