% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTest_list_nodebynode}
\alias{xmlWriteTest_list_nodebynode}
\title{Write XML Test Definition of Type 'RTestTest_list_nodebynode'}
\usage{
xmlWriteTest_list_nodebynode(elemname = "return-value",
  testname = "list_nodebynode", data = NULL, test = "absolute",
  tolerance = 0, printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The tag name of the test}

\item{testname}{(\code{character}) The testname}

\item{data}{(\code{data.frame}) The reference data for which the test should be 
written.}

\item{test}{(\code{ANY}) Named vector with test, single entry if same for all.}

\item{tolerance}{(\code{ANY}) Named vector with tolerances, single entry if same for all.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write XML Test Definition of Type 'RTestTest_list_nodebynode'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
