% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTest_variable}
\alias{xmlWriteTest_variable}
\title{Write XML Test Definition of Type 'RTestTest_variable'}
\usage{
xmlWriteTest_variable(elemname = "return-value", testname = "variable",
  test = "absolute", tolerance = 0, printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The tag name of the test}

\item{testname}{(\code{character}) The testname}

\item{test}{(\code{ANY}) Named vector with test, single entry if same for all.}

\item{tolerance}{(\code{ANY}) Named vector with tolerances, single entry if same for all.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write XML Test Definition of Type 'RTestTest_variable'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
