% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTest_vector_elementbyelement}
\alias{xmlWriteTest_vector_elementbyelement}
\title{Write XML Test Definition of Type 'RTestTest_vector_elementbyelement'}
\usage{
xmlWriteTest_vector_elementbyelement(elemname = "return-value",
  testname = "vector_elementbyelement", data = NULL,
  test = "absolute", tolerance = 0, printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The tag name of the test}

\item{testname}{(\code{character}) The testname}

\item{data}{(\code{data.frame}) The reference data for which the test should be 
written.}

\item{test}{(\code{ANY}) Named vector with test, single entry if same for all.}

\item{tolerance}{(\code{ANY}) Named vector with tolerances, single entry if same for all.}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write XML Test Definition of Type 'RTestTest_vector_elementbyelement'
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
