% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTestFunction}
\alias{xmlWriteTestFunction}
\title{Write the Test section for a RTestCase}
\usage{
xmlWriteTestFunction(elemname, testdesc = NA, execresid = NA,
  specid = NA, riskid = NA, params = "", reference = "",
  testspec = "", printXML = TRUE)
}
\arguments{
\item{elemname}{(\code{character}) The name of the element's root tag}

\item{testdesc}{(\code{character}) The description of the test's root tag}

\item{execresid}{(\code{character}) Executed Risk ID}

\item{specid}{(\code{character}) The Specification ID}

\item{riskid}{(\code{character}) The Risk ID}

\item{params}{(\code{ANY}) The Parameters of the function}

\item{reference}{(\code{ANY}) The Reference tested against}

\item{testspec}{(\code{ANY}) The test specification do calculate}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write the Test section for a RTestCase
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
