% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.xmlWrite.R
\name{xmlWriteTests}
\alias{xmlWriteTests}
\title{Write the Tests section for a RTestCase}
\usage{
xmlWriteTests(..., printXML = TRUE)
}
\arguments{
\item{...}{(\code{character}) Stuff to include in the tests section}

\item{printXML}{(\code{logical}) Print output or return xml as R object}
}
\value{
(\code{character})
}
\description{
Write the Tests section for a RTestCase
}
\author{
Matthias Pfeifer \email{matthias.pfeifer@roche.com}
}
