% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{body_cluster_function}
\alias{body_cluster_function}
\title{body_cluster_function}
\usage{
body_cluster_function(func_ptr, regionRef, fork_cluster = F)
}
\arguments{
\item{func_ptr}{Pointer to function}

\item{regionRef}{OTF2_regionRef - regionRef for function}

\item{fork_cluster}{Boolean - True if fork cluster, else psock}
}
\value{
type(as.call) Updated function body
}
\description{
Returns instrumented body for functions of type fork/psock
}
