% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{create_otf2_event}
\alias{create_otf2_event}
\title{create_otf2_event}
\usage{
create_otf2_event(func_name)
}
\arguments{
\item{func_name}{String - Name of function}
}
\value{
regionRef Int - Index of stringRef for function
}
\description{
Creates stringRef and regionRef for func_name
}
