% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evtWriter_MeasurementOnOff_client}
\alias{evtWriter_MeasurementOnOff_client}
\title{Send message to enable or disable event measurement from client side}
\usage{
evtWriter_MeasurementOnOff_client(measurementMode)
}
\arguments{
\item{measurementMode}{True to enable, else disable}
}
\value{
R_NilValue
}
\description{
Send message to enable or disable event measurement from client side
}
