% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_ll.R
\name{get_end_fork_wrapper_expression}
\alias{get_end_fork_wrapper_expression}
\title{get_end_fork_wrapper_expression}
\usage{
get_end_fork_wrapper_expression()
}
\description{
Returns wrapper expression
}
