% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fork_functions.R
\name{get_new_function_body}
\alias{get_new_function_body}
\title{get_new_function_body}
\usage{
get_new_function_body(func_ptr, func_name, regionRef)
}
\arguments{
\item{func_ptr}{Pointer to function}

\item{func_name}{Char[] - Name of function, only used for debugging}

\item{regionRef}{Int - OTF2 regionRef for function}
}
\value{
type(as.call) Updated function body
}
\description{
Returns instrumented body for function type (fork, end fork, default)
}
