% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_exception_list.R
\name{get_package_exception_list}
\alias{get_package_exception_list}
\title{get_package_exception_list
Get list of packages to not instrument}
\usage{
get_package_exception_list()
}
\value{
package_exception_list String[] List of packages to not instrument
}
\description{
get_package_exception_list
Get list of packages to not instrument
}
