% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{globalDefWriter_WriteRegion}
\alias{globalDefWriter_WriteRegion}
\title{Define new region description in global writer}
\usage{
globalDefWriter_WriteRegion(regionRef, stringRef_name)
}
\arguments{
\item{regionRef}{id/index for string}

\item{stringRef_name}{Name to be associated with region}
}
\value{
R_NilValue
}
\description{
Define new region description in global writer
}
