% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_otf2_logger}
\alias{init_otf2_logger}
\title{Fork and initialize zeromq sockets for writing globalDef definitions}
\usage{
init_otf2_logger(
  max_nprocs,
  archivePath,
  archiveName,
  collect_metrics,
  flag_print_pids
)
}
\arguments{
\item{max_nprocs}{Maximum number of R processes (ie evtWriters required)}

\item{archivePath}{Path to otf2 archive}

\item{archiveName}{Name of otf2 archive}

\item{collect_metrics}{Collect HWPC metrics via pmpmeas}

\item{flag_print_pids}{True to print pids of parent and child procs}
}
\value{
<0 if error, 0 if R master, else >0 if child
}
\description{
Fork and initialize zeromq sockets for writing globalDef definitions
}
