% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init_otf2_logger}
\alias{init_otf2_logger}
\title{Fork and initialize zeromq sockets for writing globalDef definitions}
\usage{
init_otf2_logger()
}
\value{
R_NilValue
}
\description{
Fork and initialize zeromq sockets for writing globalDef definitions
}
