% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{insert_instrumentation}
\alias{insert_instrumentation}
\title{insert_instrumentation
Insert instrumentation prefix for package_name:::func}
\usage{
insert_instrumentation(
  func,
  func_name,
  func_index,
  regionRef,
  package_name,
  flag_user_function = FALSE,
  env_is_locked = TRUE
)
}
\arguments{
\item{func}{Object - Pointer to function closure to update}

\item{func_name}{String - function name}

\item{func_index}{Integer - index of function in function_list in PROFILE_INSTRUMENTATION_DF}

\item{regionRef}{Integer - OTF2 regionRef index}

\item{package_name}{String - package name}

\item{flag_user_function}{Boolean - True if instrumenting user functoin}

\item{env_is_locked}{Boolean - True if function name-/package- space is locked}
}
\description{
insert_instrumentation
Insert instrumentation prefix for package_name:::func
}
