% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{instrument_all_functions}
\alias{instrument_all_functions}
\title{instrument_all_functions}
\usage{
instrument_all_functions(
  package_list = NULL,
  flag_user_functions = TRUE,
  flag_debug = FALSE
)
}
\arguments{
\item{package_list}{String[] - Array of package names to instrument, if none instrument all packages}

\item{flag_user_functions}{Boolean - TRUE if also flagging user functions}

\item{flag_debug}{Boolean - Enable debug statements}
}
\description{
Instrument all functions
}
