% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_ll.R
\name{instrument_user_functions}
\alias{instrument_user_functions}
\title{instrument_user_functions}
\usage{
instrument_user_functions(flag_debug = FALSE, flag_slave_proc = FALSE)
}
\arguments{
\item{flag_debug}{Boolean - Enable debug statements}

\item{flag_slave_proc}{Boolean - Enable if running on slave proc}
}
\description{
Instrument user functions
}
