% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{instrument_user_functions}
\alias{instrument_user_functions}
\title{instrument_user_functions
Instrument user functions}
\usage{
instrument_user_functions(flag_debug = FALSE)
}
\arguments{
\item{flag_debug}{Boolean - Enable debug statements}
}
\description{
instrument_user_functions
Instrument user functions
}
