% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{instrumentation_disable}
\alias{instrumentation_disable}
\title{instrumentation_disable
Disable instrumentation}
\usage{
instrumentation_disable()
}
\description{
instrumentation_disable
Disable instrumentation
}
