% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{instrumentation_enable}
\alias{instrumentation_enable}
\title{instrumentation_enable
Enable instrumentation and reset function depth}
\usage{
instrumentation_enable()
}
\description{
instrumentation_enable
Enable instrumentation and reset function depth
}
