% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_hl.R
\name{instrumentation_finalize}
\alias{instrumentation_finalize}
\title{instrumentation_finalize}
\usage{
instrumentation_finalize()
}
\description{
Close otf2 objs for instrumentation

Close otf2 objs for instrumentation
}
