% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_hl.R
\name{instrumentation_init}
\alias{instrumentation_init}
\title{instrumentation_init}
\usage{
instrumentation_init(
  flag_user_functions = T,
  collect_metrics = F,
  verbose_wrapping = F
)
}
\arguments{
\item{flag_user_functions}{Boolean - TRUE to include user functions in dataframe}

\item{collect_metrics}{Boolean - Enable papi/perf metric collection with instrumentation}

\item{verbose_wrapping}{Boolean - Print info about skipping or instrumenting each function. Produces large amount of info to stdout. Intended for developers}
}
\description{
Create otf2 objs for instrumentation, and initiate global vars
}
