% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{instrumentation_init}
\alias{instrumentation_init}
\title{instrumentation_init}
\usage{
instrumentation_init(r_profiling = T, verbose_wrapping = F)
}
\arguments{
\item{r_profiling}{Boolean - True to enable R-based eventlogger and runtime info}

\item{verbose_wrapping}{Boolean - Print info about skipping or instrumenting each function. Produces large amount of info to stdout}
}
\description{
Create otf2 objs for instrumentation, and initiate global vars
}
