% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{instrumentation_init}
\alias{instrumentation_init}
\title{instrumentation_init}
\usage{
instrumentation_init(
  r_profiling = T,
  flag_user_functions = T,
  verbose_wrapping = F
)
}
\arguments{
\item{r_profiling}{Boolean - TRUE to enable R-based eventlogger and runtime info}

\item{flag_user_functions}{Boolean - TRUE to include user functions in dataframe}

\item{verbose_wrapping}{Boolean - Print info about skipping or instrumenting each function. Produces large amount of info to stdout}
}
\description{
Create otf2 objs for instrumentation, and initiate global vars
}
