% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument.R
\name{print_function_from_index}
\alias{print_function_from_index}
\title{print_function_from_index}
\usage{
print_function_from_index(func_indexes)
}
\arguments{
\item{func_indexes}{Int[]}
}
\description{
Prints names of given functions from indexes, only intended for debugging
}
