% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{print_instrumentation}
\alias{print_instrumentation}
\title{print_instrumentation
Print table of reduce dataframe}
\usage{
print_instrumentation(flag_debug = TRUE)
}
\arguments{
\item{flag_debug}{Boolean - Enable debug header}
}
\description{
print_instrumentation
Print table of reduce dataframe
}
