% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{reduce_dataframe}
\alias{reduce_dataframe}
\title{reduce_dataframe}
\usage{
reduce_dataframe(df = pkg.env$PROFILE_INSTRUMENTATION_DF)
}
\arguments{
\item{df}{Dataframe - Created by create_dataframe() for collecting tracing info}
}
\value{
reduced_df Dataframe - All entries of functions which were called
}
\description{
Reduce dataframe to non-zero function_count entries
}
