% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{reduce_dataframe}
\alias{reduce_dataframe}
\title{reduce_dataframe}
\usage{
reduce_dataframe(df)
}
\arguments{
\item{df}{Dataframe - Created by create_dataframe() for collecting tracing info}
}
\description{
Reset function_count and function_time info to zero
}
