% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_event_instrumentation_loggers.R
\name{reset_dataframe}
\alias{reset_dataframe}
\title{reset_dataframe}
\usage{
reset_dataframe(df = pkg.env$PROFILE_INSTRUMENTATION_DF)
}
\arguments{
\item{df}{Dataframe - Created by create_dataframe() for collecting tracing info}
}
\description{
Reset function_count and function_time info to zero
}
