% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_evtWriter.R
\name{rTest_init_EvtWriter}
\alias{rTest_init_EvtWriter}
\title{rTest_init_ - doxygen documentation
Simple wrapper for C++ otf2 init function}
\usage{
rTest_init_EvtWriter(flag_debug = F)
}
\arguments{
\item{flag_debug}{Boolean - True to enable verbose stdout}
}
\description{
rTest_init_ - doxygen documentation
Simple wrapper for C++ otf2 init function
}
