% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otf2_globalDefWriter.R
\name{rTest_init_GlobalDefWriter}
\alias{rTest_init_GlobalDefWriter}
\title{init_GlobalDefWriter - doxygen documentation
More info}
\usage{
rTest_init_GlobalDefWriter()
}
\description{
init_GlobalDefWriter - doxygen documentation
More info
}
