% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_instrument_ll.R
\name{total_num_functions}
\alias{total_num_functions}
\title{total_num_functions}
\usage{
total_num_functions(flag_user_functions = FALSE, debug_flag = FALSE)
}
\arguments{
\item{flag_user_functions}{Boolean - TRUE if also flagging user functions}

\item{debug_flag}{Boolean - TRUE to print debug information}
}
\value{
Int[] Number of functions per package
}
\description{
Find total number of loaded packages and functions
}
